% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\name{abundanceBoxplot,MultiAssayExperiment-method}
\alias{abundanceBoxplot,MultiAssayExperiment-method}
\title{abundanceBoxplot}
\usage{
\S4method{abundanceBoxplot}{MultiAssayExperiment}(object, dataType, pngFile = NULL)
}
\arguments{
\item{dataType}{omic data type}

\item{MultiAssayExperiment}{an object of Class MultiAssayExperiment}
}
\description{
abundanceBoxplot
}
