% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\docType{methods}
\name{barplotPCAnorm,FlomicsExperiment-method}
\alias{barplotPCAnorm,FlomicsExperiment-method}
\title{barplotPCAnorm}
\usage{
\S4method{barplotPCAnorm}{FlomicsExperiment}(object,
  condition = "samples")
}
\arguments{
\item{condition}{}

\item{FlomicsExperiment}{}

\item{colors}{color palette}
}
\description{
barplotPCAnorm
}
