% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\name{barplotPCAnorm,MultiAssayExperiment-method}
\alias{barplotPCAnorm,MultiAssayExperiment-method}
\title{barplotPCAnorm}
\usage{
\S4method{barplotPCAnorm}{MultiAssayExperiment}(object, condition = "samples")
}
\arguments{
\item{colors}{color palette}
}
\value{

}
\description{
barplotPCAnorm
}
