% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\docType{methods}
\name{boxplotQCnorm,FlomicsExperiment-method}
\alias{boxplotQCnorm,FlomicsExperiment-method}
\title{boxplotQCnorm}
\usage{
\S4method{boxplotQCnorm}{FlomicsExperiment}(object)
}
\arguments{
\item{FlomicsExperiment}{an object of Class FlomicsExperiment}
}
\description{
boxplotQCnorm
}
