% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_01_load_data.R
\name{checkExpDesignCompleteness,RflomicsSE-method}
\alias{checkExpDesignCompleteness,RflomicsSE-method}
\alias{checkExpDesignCompleteness}
\alias{checkExpDesignCompleteness,RflomicsMAE-method}
\title{checkExpDesignCompleteness}
\usage{
\S4method{checkExpDesignCompleteness}{RflomicsSE}(object, sampleList = NULL)

\S4method{checkExpDesignCompleteness}{RflomicsMAE}(object, omicName, sampleList = NULL)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE-class}}

\item{sampleList}{list of samples to check.}

\item{omicName}{the name of the data the normalization have to be applied to.}
}
\value{
a string with message
}
\description{
This method checks some experimental design characteristics.
 A complete design (all combinations of factor modalities with at least 2 replicates for each have to be present) with
 at least one biological and one batch factors are required to use the RFLOMICS workflow.
}
