% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_02_set_stat_model.R,
%   R/utils_02_set_stat_model.R
\name{getSelectedContrasts,RflomicsMAE-method}
\alias{getSelectedContrasts,RflomicsMAE-method}
\alias{getSelectedContrasts}
\alias{getSelectedContrasts,RflomicsSE-method}
\alias{getPossibleContrasts}
\title{Get selected contrasts for the differential analysis}
\usage{
\S4method{getSelectedContrasts}{RflomicsMAE}(object)

\S4method{getSelectedContrasts}{RflomicsSE}(object)

getPossibleContrasts(
  object,
  formula = object@metadata$design$Model.formula,
  typeContrast = c("simple", "averaged", "interaction"),
  modalities = NULL,
  returnTable = FALSE
)
}
\arguments{
\item{object}{a MAE object (produced by Flomics) or a summarized 
experiment produced by Flomics after a differential analysis.}

\item{formula}{The formula used to compute all possible contrasts. 
Default is the formula found in the object.}

\item{typeContrast}{the type of contrast from which the possible 
contrasts are extracted. Default is all contrasts types.}

\item{modalities}{specific levels for the contrast selection}

\item{returnTable}{return a dataTable with all contrasts information}
}
\value{
a dataTable

a dataTable

a character vector or a dataTable
}
\description{
Get selected contrasts for the differential analysis

Get selected contrasts for the differential analysis

Get selected contrasts for the differential analysis
}
