% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Class.R
\docType{class}
\name{DiffAnalysis-class}
\alias{DiffAnalysis-class}
\alias{.DiffAnalysis}
\title{DiffAnalysis class}
\value{

}
\description{
Store the results of a differential analysis
}
\section{Slots}{

\describe{
\item{\code{method}}{The name of the Differential method applied}

\item{\code{ListOfDEResults}}{A list of data.frame with the following columns:
\describe{
\item{id}{feature identifier}
\item{BaseMean}{}
\item{BaseMeanCondA}{}
\item{BaseMeanCondB}{}
\item{logFC}{}
\item{pvalue}{}
\item{FDR}{}
}}

\item{\code{ListOfRawMethodResults}}{A list of objects returned by the used method (ex: list of object of class DGELRT)}
}}

