% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\name{DiffAnalysis-Constructor}
\alias{DiffAnalysis-Constructor}
\alias{DiffAnalysis}
\title{[\code{\link{DiffAnalysis-class}}] Class constructor}
\usage{
DiffAnalysis()
}
\value{
An object of class [\code{\link{DiffAnalysis-class}}]
}
\description{

}
