% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{edgeR.AnaDiff}
\alias{edgeR.AnaDiff}
\title{edgeR.AnaDiff}
\usage{
edgeR.AnaDiff(object, data, FDR, clustermq)
}
\arguments{
\item{data}{Omic data type}

\item{FDR}{The false discovery rate to apply}

\item{clustermq}{A boolean indicating if the constrasts have to be computed in local or in a distant machine}
}
\value{

}
\description{
edgeR.AnaDiff
}
