% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_examples.R
\name{initExampleMAE}
\alias{initExampleMAE}
\alias{generateExample}
\title{Initializing an Example MultiAssayExperiment compatible with RFLOMICS
 functions.}
\usage{
initExampleMAE()

generateExample(
  processing = TRUE,
  diffanalysis = TRUE,
  coexp = TRUE,
  annotation = TRUE,
  integration = TRUE
)
}
\arguments{
\item{processing}{boolean. If TRUE, processing step is done.}

\item{diffanalysis}{boolean. If TRUE, differential analysis step is done.}

\item{coexp}{boolean. If TRUE, co-expression step is done.}

\item{annotation}{boolean. If TRUE, enrichment analysis step is done.}

\item{integration}{boolean. If TRUE, integration step is done.}
}
\value{
a multiAssayExperiment ready for use in RFLOMICS

a RflomicsMAE object, results of some RFLOMICS analyses
}
\description{
Initializing an Example MultiAssayExperiment compatible with RFLOMICS
 functions.

generateExample
}
\examples{
initExampleMAE()

MAE <- generateExample(coexp = FALSE, annotation = FALSE, integration = FALSE)

}
