% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Class.R
\docType{class}
\name{ExpDesign-class}
\alias{ExpDesign-class}
\alias{.ExpDesign}
\title{ExpDesign Class}
\value{
ExpDesign object
}
\description{
ExpDesign Class
}
\section{Slots}{

\describe{
\item{\code{List.Factors}}{list.}

\item{\code{Factors.Type}}{vector.}

\item{\code{Model.formula}}{formula.}

\item{\code{Contrasts.List}}{list.}
}}

