% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\name{ExpDesign-Constructor}
\alias{ExpDesign-Constructor}
\alias{ExperimentalDesign}
\title{[\code{\link{ExpDesign-class}}] Class constructor}
\usage{
ExperimentalDesign(ExpDesign)
}
\arguments{
\item{dF.List}{A list of factor}
}
\value{
An object of class [\code{\link{ExpDesign-class}}]
}
\description{
[\code{\link{ExpDesign-class}}] Class constructor
}
