% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_04_diff_analysis.R
\name{filterDiffAnalysis,RflomicsSE-method}
\alias{filterDiffAnalysis,RflomicsSE-method}
\alias{filterDiffAnalysis}
\alias{filterDiffAnalysis,RflomicsMAE-method}
\title{filterDiffAnalysis
Filter differential analysis}
\usage{
\S4method{filterDiffAnalysis}{RflomicsSE}(object, p.adj.cutoff = NULL, logFC.cutoff = NULL)

\S4method{filterDiffAnalysis}{RflomicsMAE}(object, SE.name, p.adj.cutoff = NULL, logFC.cutoff = NULL)
}
\arguments{
\item{object}{A RflomicsSE object}

\item{p.adj.cutoff}{adjusted pvalue cutoff. Default is the parameter from 
the differential analysis.}

\item{logFC.cutoff}{cutoff for absolute value of log2FC. Default is the 
parameter from the differential analysis.}

\item{SE.name}{the name of the data to fetch in the object if the object 
is a \link{RflomicsMAE}}
}
\value{
A \link{RflomicsSE} object or a \link{RflomicsMAE}, depending on the object type, 
where the differential analysis results have been actualized with the 
new parameters.
}
\description{
filterDiffAnalysis
Filter differential analysis

filterDiffAnalysis
}
\examples{
# Set the data path
datPath <- paste0(system.file(package = "RFLOMICS"), "/ExamplesFiles/ecoseed/")

# Load the experimental design
ExpDesign <- readExpDesign(file = paste0(datPath, "condition.txt"))

# Set factor name, levels, level of reference and type
facRef <- data.frame(factorName   = c("Repeat", "temperature" , "imbibition"), 
                     factorRef = c("rep1", "Low", "DS"), 
                     factorType = c("batch",  "Bio", "Bio"), 
                     factorLevels = c("rep1,rep2,rep3","Low,Medium,Elevated","DS,EI,LI"))

# Load the omics data
omicsData <- list( 
readOmicsData(file = paste0( datPath, "transcriptome_ecoseed.txt")), 
readOmicsData(file = paste0(datPath, "proteome_ecoseed.txt")))
 
# Instantiate an object of class RflomicsMAE 
MAE <- createRflomicsMAE( projectName = "Tests", omicsData   = omicsData,
                         omicsNames  = c("RNAtest", "protetest"),
                         omicsTypes  = c("RNAseq","proteomics"),
                         ExpDesign   = ExpDesign, factorRef   = facRef)
names(MAE) <- c("RNAtest", "protetest")

# Set the statistical model and contrasts to test
formulae <- generateModelFormulae(MAE)
MAE <- setModelFormula(MAE, formulae[[1]])  

# Get the contrasts List and choose the first 3 contrasts of type averaged
contrastList <- getPossibleContrasts(MAE, formula = formulae[[1]], 
                                     typeContrast = "averaged", 
                                     returnTable = TRUE)[c(1, 2, 3),]

# Run the data preprocessing and perform the differential analysis
MAE |>  runTransformData(SE.name = "protetest",  transformMethod = "log2") |>
 filterLowAbundance(SE.name = "RNAtest")                           |>    
 runNormalization(SE.name = "RNAtest",   normMethod = "TMM")       |>
 runNormalization(SE.name = "protetest", normMethod = "median")    |>
 runDiffAnalysis(SE.name = "protetest", method = "limmalmFit", 
                 contrastList = contrastList)  |>
 runDiffAnalysis(SE.name = "RNAtest", method = "edgeRglmfit", 
                 contrastList = contrastList) |> 
 filterDiffAnalysis(SE.name = "protetest", p.adj.cutoff = 0.01, logFC.cutoff = 1) |> 
 filterDiffAnalysis(SE.name = "RNAtest", p.adj.cutoff = 0.01, logFC.cutoff = 1)
}
