% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\docType{methods}
\name{FilterLowAbundance,FlomicsExperiment-method}
\alias{FilterLowAbundance,FlomicsExperiment-method}
\title{FilterLowAbundance}
\usage{
\S4method{FilterLowAbundance}{FlomicsExperiment}(object, threshold)
}
\arguments{
\item{threshold}{}

\item{FlomicsExperiment}{}
}
\value{
FlomicsExperiment
}
\description{
FilterLowAbundance
}
