% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\name{FilterLowAbundance,MultiAssayExperiment-method}
\alias{FilterLowAbundance,MultiAssayExperiment-method}
\title{FilterLowAbundance}
\usage{
\S4method{FilterLowAbundance}{MultiAssayExperiment}(
  object,
  data,
  Filter_Strategy = "NbConditions",
  CPM_Cutoff = 5
)
}
\arguments{
\item{threshold}{}
}
\value{
MultiAssayExperiment
}
\description{
FilterLowAbundance
}
