% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_03_data_explor.R
\name{filterLowAbundance,RflomicsSE-method}
\alias{filterLowAbundance,RflomicsSE-method}
\alias{filterLowAbundance}
\alias{filterLowAbundance,RflomicsMAE-method}
\title{filterLowAbundance}
\usage{
\S4method{filterLowAbundance}{RflomicsSE}(
  object,
  filterMethod = "CPM",
  filterStrategy = "NbConditions",
  cpmCutoff = 5
)

\S4method{filterLowAbundance}{RflomicsMAE}(
  object,
  SE.name,
  filterMethod = "CPM",
  filterStrategy = "NbConditions",
  cpmCutoff = 5
)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE}}

\item{filterMethod}{The filtering model ("CPM")}

\item{filterStrategy}{The filtering strategy ("NbConditions" or "NbReplicates")}

\item{cpmCutoff}{The CPM cutoff.}

\item{SE.name}{the name of the data the normalization have to be applied to.}
}
\value{
An object of class \link{RflomicsSE}
}
\description{
This function aims at removing genes/transcript from the count data matrix of an omic of type "RNAseq".
by applying filtering criterion described in reference.
By default, gene/transcript with 0 count are removed from the data. The function then
computes the count per million or read (CPM) for each gene in each sample and gives by
genes the number of sample(s) which are over the cpmCutoff (NbOfsample_over_cpm).
Then Two filtering strategies are proposed:
\itemize{
\item{NbConditions: }{keep gene if the NbOfsample_over_cpm >= NbConditions}
\item{NbReplicates: }{keep gene if the NbOfsample_over_cpm >= min(NbReplicat)}
\item{filterByExpr:} {the default filtering method implemented in the edgeR filterByExpr() function.}
}
}
\details{
Filtered dataset is stored in the ExperimentList slot of the \link{RflomicsSE} object
as a List named (DataName.filtred).
List of filtered features are stored as a named list ("FilteredFeatures") in the metadata slot of a
given data set, stored itself in the ExperimentList slot of a \link{RflomicsSE} object.
}
\examples{

# Set the data path
datPath <- paste0(system.file(package = "RFLOMICS"), "/ExamplesFiles/ecoseed/")

# Load the experimental design
ExpDesign <- readExpDesign(file = paste0(datPath, "condition.txt"))

# Set factor name, levels, level of reference and type
facRef <- data.frame(factorName   = c("Repeat", "temperature" , "imbibition"), 
                     factorRef = c("rep1", "Low", "DS"), 
                     factorType = c("batch",  "Bio", "Bio"), 
                     factorLevels = c("rep1,rep2,rep3","Low,Medium,Elevated","DS,EI,LI"))

# Load the omics data
omicsData <- list( 
readOmicsData(file = paste0( datPath, "transcriptome_ecoseed.txt")))
 
# Instantiate an object of class RflomicsMAE 
MAE <- createRflomicsMAE( projectName = "Tests", omicsData   = omicsData,
                         omicsNames  = c("RNAtest"), omicsTypes  = c("RNAseq"),
                         ExpDesign   = ExpDesign, factorRef   = facRef)
names(MAE) <- c("RNAtest")

# Set the statistical model and contrasts to test
formulae <- generateModelFormulae(MAE)
MAE <- setModelFormula(MAE, formulae[[1]])  

# Get the contrasts List and choose the first 3 contrasts of type averaged
contrastList <- getPossibleContrasts(MAE, formula = formulae[[1]], 
                                     typeContrast = "averaged", 
                                     returnTable = TRUE)[c(1),]

# Run the data preprocessing and perform the differential analysis
MAE |>  filterLowAbundance(SE.name = "RNAtest") 
}
\references{
Lambert, I., Paysant-Le Roux, C., Colella, S. et al. DiCoExpress: a tool to process multifactorial RNAseq experiments from quality controls to co-expression analysis through differential analysis based on contrasts inside GLM models. Plant Methods 16, 68 (2020).
}
\seealso{
edgeR::filterByExpr
}
