% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Class.R
\docType{class}
\name{FlomicsExperiment-class}
\alias{FlomicsExperiment-class}
\alias{.FlomicsExperiment}
\title{FlomicsExperiment-Class}
\value{
FlomicsExperiment object
}
\description{
This class extend the [\code{\link{SummarizedExperiment}}] class by adding a named vector
(colDataStruc) giving the number of factors in the RNAseq design and the number of QC parameters.
}
\section{Slots}{

\describe{
\item{\code{colDataStruc}}{giving the number of factor in the RNAseq design and the number of QC parameters.}
}}

