% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_02_set_stat_model.R
\name{generateContrastMatrix,RflomicsSE-method}
\alias{generateContrastMatrix,RflomicsSE-method}
\alias{generateContrastMatrix}
\alias{generateContrastMatrix,RflomicsMAE-method}
\title{generateContrastMatrix}
\usage{
\S4method{generateContrastMatrix}{RflomicsSE}(object, contrastList = NULL)

\S4method{generateContrastMatrix}{RflomicsMAE}(object, SE.name, contrastList = NULL)
}
\arguments{
\item{object}{An object of class \link{RflomicsMAE-class}}

\item{contrastList}{A data.frame of contrast}

\item{SE.name}{the name of the \link{RflomicsSE-class} to access.}
}
\value{
An object of class \link{RflomicsSE-class}

An object of class \link{RflomicsMAE-class}
}
\description{
Defines contrast matrix or contrast list with contrast 
name and contrast coefficients

Defines contrast matrix or contrast list with contrast name 
and contrast coefficients
}
\seealso{
generateContrastMatrix

generateContrastMatrix
}
\author{
Christine Paysant-Le Roux, adapted by Nadia Bessoltane
}
