% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_02_set_stat_model.R
\name{generateModelFormulae,RflomicsMAE-method}
\alias{generateModelFormulae,RflomicsMAE-method}
\alias{generateModelFormulae}
\title{generateModelFormulae}
\usage{
\S4method{generateModelFormulae}{RflomicsMAE}(object)
}
\arguments{
\item{object}{a RflomicsMAE object}
}
\value{
a named list of object of class formula
}
\description{
From a vector of character giving the name of the factors of an omics
 experiment, and their type of effect: biological or batch, it returns all 
 models formulae that can be formulated in association with this factors. 
 Batch effect factors do not appear in interaction terms with biological 
 factor. Model formulae stop in second order interaction.
}
