% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_00_Common.R
\name{getAnnotAnalysesSummary,RflomicsMAE-method}
\alias{getAnnotAnalysesSummary,RflomicsMAE-method}
\alias{getAnnotAnalysesSummary}
\title{getAnnotAnalysesSummary}
\usage{
\S4method{getAnnotAnalysesSummary}{RflomicsMAE}(
  object,
  from = "DiffExpEnrichAnal",
  matrixType = "presence",
  ...
)
}
\arguments{
\item{object}{An object of class \link{RflomicsMAE}. It is expected the SE 
object is produced by rflomics previous analyses, as it relies on their 
results.}

\item{from}{indicates if the enrichment results are taken from differential 
analysis results (DiffExpEnrichAnal) or from the co-expression analysis 
results (CoExpEnrichAnal)}

\item{matrixType}{Heatmap matrix to plot, one of GeneRatio, p.adjust 
or presence.}

\item{...}{more arguments for ComplexHeatmap::Heatmap.}
}
\value{
A list of heatmaps, one for each ontology/domain.
}
\description{
TODO
}
