% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_03_data_explor.R
\name{getCoeffNorm,RflomicsSE-method}
\alias{getCoeffNorm,RflomicsSE-method}
\alias{getCoeffNorm}
\alias{getCoeffNorm,RflomicsMAE-method}
\title{get Coeff Norm}
\usage{
\S4method{getCoeffNorm}{RflomicsSE}(object)

\S4method{getCoeffNorm}{RflomicsMAE}(object, SE.name)
}
\arguments{
\item{object}{of class RflomicsSE}

\item{SE.name}{the name of the data to fetch in the object if the object is a RflomicsMAE}
}
\value{
Coefficient of normalization
}
\description{
get Coeff Norm

getCoeffNorm
}
