% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_00_Common.R
\name{getCoExpAnalysesSummary,RflomicsMAE-method}
\alias{getCoExpAnalysesSummary,RflomicsMAE-method}
\alias{getCoExpAnalysesSummary}
\title{getCoExpAnalysesSummary}
\usage{
\S4method{getCoExpAnalysesSummary}{RflomicsMAE}(object, omicNames = NULL, ...)
}
\arguments{
\item{object}{An object of class \link{RflomicsMAE}. 
It is expected the SE object is produced by rflomics previous analyses, 
as it relies on their results.}

\item{omicNames}{the name of the experiment to summarize.}

\item{...}{Not in use at the moment}
}
\value{
A ggplot object
}
\description{
TODO
}
