% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_05_coexp_analysis.R
\name{getCoexpSettings,RflomicsSE-method}
\alias{getCoexpSettings,RflomicsSE-method}
\alias{getCoexpSettings}
\alias{getCoexpSettings,RflomicsMAE-method}
\title{Get co-expression analysis setting parameters}
\usage{
\S4method{getCoexpSettings}{RflomicsSE}(object)

\S4method{getCoexpSettings}{RflomicsMAE}(object, SE.name)
}
\arguments{
\item{object}{of class RflomicsSE}

\item{SE.name}{the name of the data to fetch in the object if the object 
is a RflomicsMAE}
}
\value{
List of co-expression analysis setting parameters.

list of co-expression analysis setting parameters
}
\description{
Get co-expression analysis setting parameters

getCoexpSettings
}
