% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetContrasts}
\alias{GetContrasts}
\title{GetContrasts}
\usage{
GetContrasts(design)
}
\arguments{
\item{An}{object of class design}
}
\value{
An object of class design
}
\description{
GetContrasts
}
\examples{

}
