% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_05_coexp_analysis.R
\name{getCoseqClusters,RflomicsSE-method}
\alias{getCoseqClusters,RflomicsSE-method}
\alias{getCoseqClusters}
\alias{getCoseqClusters,RflomicsMAE-method}
\title{get cluster}
\usage{
\S4method{getCoseqClusters}{RflomicsSE}(object)

\S4method{getCoseqClusters}{RflomicsMAE}(object, SE.name)
}
\arguments{
\item{object}{an oblject of class RflomicsMAE, produced by rflomics}

\item{SE.name}{the name of the data to fetch in the object if the object is 
a RflomicsMAE}
}
\value{
all clusters

all clusters
}
\description{
get cluster

get members of a cluster
}
