% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_04_diff_analysis.R
\name{getDEList,RflomicsSE-method}
\alias{getDEList,RflomicsSE-method}
\alias{getDEList}
\alias{getDEList,RflomicsMAE-method}
\title{Operation on differential analyses lists. Get union vector of DE 
entities from list of contrasts}
\usage{
\S4method{getDEList}{RflomicsSE}(object, contrasts = NULL, operation = "union")

\S4method{getDEList}{RflomicsMAE}(object, SE.name, contrasts = NULL, operation = "union")
}
\arguments{
\item{object}{an object of class \link{RflomicsSE}. Expects to find a slot with 
differential analyses results.}

\item{contrasts}{Vector of characters, expect to be contrast names. 
Default is null, the operation (union) is performed
on every contrasts found.}

\item{operation}{character. Either union or intersection.
Defines the operation to perform on the DE lists from the contrasts.}

\item{SE.name}{the name of the data to fetch in the object if the object is 
a RflomicsMAE}
}
\value{
vector of unique DE entities
}
\description{
Operation on differential analyses lists. Get union vector of DE 
entities from list of contrasts

getDEList
}
\seealso{
[runDiffAnalysis()], [filterDiffAnalysis()]
}
