% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_04_diff_analysis.R
\name{getDEMatrix,RflomicsSE-method}
\alias{getDEMatrix,RflomicsSE-method}
\alias{getDEMatrix}
\alias{getDEMatrix,RflomicsMAE-method}
\title{Get DE matrix}
\usage{
\S4method{getDEMatrix}{RflomicsSE}(object)

\S4method{getDEMatrix}{RflomicsMAE}(object, SE.name)
}
\arguments{
\item{object}{of class \link{RflomicsSE} or \link{RflomicsMAE}}

\item{SE.name}{the name of the data to fetch in the object if the object is  \link{RflomicsMAE}}
}
\value{
a matrix of results from the differential analyses.
}
\description{
Get DE matrix

getDEMatrix
}
