% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetDesignFromNames}
\alias{GetDesignFromNames}
\title{GetDesignFromNames}
\usage{
GetDesignFromNames(samples_name)
}
\arguments{
\item{samples_name}{a vector of sample names giving the designs factor, each
separated by "_"}
}
\value{
a named list of two elements
\itemize{
\item nb_dfac: the number of design factors
\item tmpDesign: a data frame with the names of design factors in columns (ie: dFac1) and all
the modalities in row
}
}
\description{
GetDesignFromNames
}
\examples{


}
