% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_00_Common.R
\name{getDiffAnalysesSummary,RflomicsMAE-method}
\alias{getDiffAnalysesSummary,RflomicsMAE-method}
\alias{getDiffAnalysesSummary}
\title{getDiffAnalysesSummary}
\usage{
\S4method{getDiffAnalysesSummary}{RflomicsMAE}(object, plot = FALSE)
}
\arguments{
\item{object}{An object of class \link{RflomicsMAE-class}}

\item{plot}{FALSE or TRUE}
}
\value{
a data.frame with differential analyses summary
}
\description{
A short description...
}
