% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_06_annot_enrichment.R
\name{getEnrichPvalue,RflomicsSE-method}
\alias{getEnrichPvalue,RflomicsSE-method}
\alias{getEnrichPvalue}
\title{Get the pvalue threshold used in enrichment analysis}
\usage{
\S4method{getEnrichPvalue}{RflomicsSE}(object, from = "DiffExpEnrichAnal", database = "GO")
}
\arguments{
\item{object}{a RflomicsSE object}

\item{from}{where to search for the results (either coexp or diffExp)}

\item{database}{which database (GO, KEGG, custom...)}
}
\value{
the pvalue cutoff used for the analysis.
}
\description{
Get the pvalue threshold used in enrichment analysis
}
\examples{
# Generate RflomicsMAE for example
MAEtest <- generateExample(annotation = FALSE, integration = FALSE)
# Run KEGG annotation on differential analysis results
MAEtest[["protetest"]] <- runAnnotationEnrichment(MAEtest[["protetest"]],
                        list_args = list(organism = "ath",
                                         keyType = "kegg",
                                         pvalueCutoff = 0.1),
                        from = "DiffExp", database = "KEGG")
# Search for the pvalue cutoff:
getEnrichPvalue(MAEtest[["protetest"]], from = "diffexp", database = "KEGG")

}
