% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_06_annot_enrichment.R
\name{getEnrichRes,RflomicsSE-method}
\alias{getEnrichRes,RflomicsSE-method}
\alias{getEnrichRes}
\alias{getEnrichRes,RflomicsMAE-method}
\title{Get a particular enrichment result}
\usage{
\S4method{getEnrichRes}{RflomicsSE}(
  object,
  contrastName = NULL,
  from = "DiffExpEnrichAnal",
  database = "GO",
  domain = NULL
)

\S4method{getEnrichRes}{RflomicsMAE}(
  object,
  experiment,
  contrastName = NULL,
  from = "DiffExpEnrichAnal",
  database = "GO",
  domain = NULL
)
}
\arguments{
\item{object}{a RflomicsSE object or a RflomicsMAE object. Should have at
least results from an enrichment, on differential analysis or co expression
clusters.}

\item{contrastName}{the contrast or cluster name on which the enrichment
was perform.}

\item{from}{either diffexp or coexp, indicates where to search for the
results}

\item{database}{the database used for the enrichment (GO, KEGG or custom)}

\item{domain}{the subonology or subdomain for the database (eg CC, MF or
BP for GO.)}

\item{experiment}{if the object is a RflomicsMAE, then experiment is the
name of the RflomicsSE to look for.}
}
\value{
enrichment results given in the form of lists of clusterprofiler
results.
}
\description{
Get a particular enrichment result
}
\examples{
# Generate RflomicsMAE for example
MAEtest <- generateExample(annotation = FALSE, integration = FALSE)
# Run KEGG annotation on differential analysis results
MAEtest[["protetest"]] <- runAnnotationEnrichment(MAEtest[["protetest"]],
                        list_args = list(organism = "ath",
                                         keyType = "kegg",
                                         pvalueCutoff = 0.1),
                        from = "DiffExp", database = "KEGG")
# Get all results from KEGG on differential expression lists:
getEnrichRes(MAEtest[["protetest"]],
              from = "diffexp", database = "KEGG")
}
