% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_03_data_explor.R
\name{getFilteredFeatures,RflomicsSE-method}
\alias{getFilteredFeatures,RflomicsSE-method}
\alias{getFilteredFeatures}
\alias{getFilteredFeatures,RflomicsMAE-method}
\title{get Filtered Features}
\usage{
\S4method{getFilteredFeatures}{RflomicsSE}(object)

\S4method{getFilteredFeatures}{RflomicsMAE}(object, SE.name)
}
\arguments{
\item{object}{of class RflomicsSE}

\item{SE.name}{the name of the data to fetch in the object if the 
object is a RflomicsMAE}
}
\value{
List of differential analysis setting parametres.
}
\description{
get Filtered Features

getFilteredFeatures
}
