% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_03_data_explor.R
\name{getFilterSettings,RflomicsSE-method}
\alias{getFilterSettings,RflomicsSE-method}
\alias{getFilterSettings}
\alias{getFilterSettings,RflomicsMAE-method}
\title{get Filter setting parameters}
\usage{
\S4method{getFilterSettings}{RflomicsSE}(object)

\S4method{getFilterSettings}{RflomicsMAE}(object, SE.name)
}
\arguments{
\item{object}{of class RflomicsSE}

\item{SE.name}{the name of the data to fetch in the object if the object is 
a RflomicsMAE}
}
\value{
List of differential analysis setting parametres.
}
\description{
get Filter setting parameters

getFilterSettings
}
