% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_02_set_stat_model.R
\name{getModelFormula,RflomicsMAE-method}
\alias{getModelFormula,RflomicsMAE-method}
\alias{getModelFormula}
\alias{getModelFormula,RflomicsSE-method}
\title{getModelFormula}
\usage{
\S4method{getModelFormula}{RflomicsMAE}(object)

\S4method{getModelFormula}{RflomicsSE}(object)
}
\arguments{
\item{object}{a RflomicsSE}
}
\value{
a formula
}
\description{
Get model formula from a \code{RflomicsMAE} or \code{RflomicsSE} object
}
\seealso{
\link{RflomicsMAE-methods}
}
