% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetModelFormulae}
\alias{GetModelFormulae}
\title{GetModelFormulae}
\usage{
GetModelFormulae(Factors.Name, Factors.Type = NULL)
}
\arguments{
\item{Factors.Type}{}
}
\value{
a formulae
}
\description{
GetModelFormulae
}
\examples{

}
