% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_03_data_explor.R
\name{getNormSettings,RflomicsSE-method}
\alias{getNormSettings,RflomicsSE-method}
\alias{getNormSettings}
\alias{getNormSettings,RflomicsMAE-method}
\title{Get normalization parameters}
\usage{
\S4method{getNormSettings}{RflomicsSE}(object)

\S4method{getNormSettings}{RflomicsMAE}(object, SE.name)
}
\arguments{
\item{object}{of class RflomicsSE}

\item{SE.name}{the name of the data to fetch in the object if the object 
is a RflomicsMAE}
}
\value{
List of differential analysis setting parametres.
}
\description{
Get normalization parameters

getNormSettings
}
