% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_03_data_explor.R
\name{getTransSettings,RflomicsSE-method}
\alias{getTransSettings,RflomicsSE-method}
\alias{getTransSettings}
\alias{getTransSettings,RflomicsMAE-method}
\title{get transformation  parameters}
\usage{
\S4method{getTransSettings}{RflomicsSE}(object)

\S4method{getTransSettings}{RflomicsMAE}(object, SE.name)
}
\arguments{
\item{object}{of class RflomicsSE}

\item{SE.name}{the name of the data to fetch in the object if the object 
is a RflomicsMAE}
}
\value{
List of transformation parametres.
}
\description{
get transformation  parameters

getTransSettings
}
