% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_02_set_stat_model.R
\name{getValidContrasts,RflomicsMAE-method}
\alias{getValidContrasts,RflomicsMAE-method}
\alias{getValidContrasts}
\alias{getValidContrasts,RflomicsSE-method}
\title{getValidContrasts}
\usage{
\S4method{getValidContrasts}{RflomicsMAE}(object, omicName)

\S4method{getValidContrasts}{RflomicsSE}(object)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE-class}}

\item{omicName}{a dataset name}
}
\value{
An object of class \link{RflomicsMAE-class}

An object of class \link{RflomicsSE-class}
}
\description{
Set Valid Contrasts

Set Valid Contrasts
}
