% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_07_Integration.R
\name{getMixOmics,RflomicsMAE-method}
\alias{getMixOmics,RflomicsMAE-method}
\alias{getMixOmics}
\alias{getMOFA,RflomicsMAE-method}
\alias{getMOFA}
\alias{getMOFASettings,RflomicsMAE-method}
\alias{getMOFASettings}
\alias{getMixOmicsSettings,RflomicsMAE-method}
\alias{getMixOmicsSettings}
\alias{setMOFA,RflomicsMAE-method}
\alias{setMOFA}
\alias{setMixOmics,RflomicsMAE-method}
\alias{setMixOmics}
\alias{sumMixOmics,RflomicsMAE-method}
\alias{sumMixOmics}
\title{Get a particular multi-omics result or settings.}
\usage{
\S4method{getMixOmics}{RflomicsMAE}(object, response = NULL, onlyResults = TRUE)

\S4method{getMOFA}{RflomicsMAE}(object, onlyResults = TRUE)

\S4method{getMOFASettings}{RflomicsMAE}(object)

\S4method{getMixOmicsSettings}{RflomicsMAE}(object)

\S4method{setMOFA}{RflomicsMAE}(object, results = NULL)

\S4method{setMixOmics}{RflomicsMAE}(object, results = NULL)

\S4method{sumMixOmics}{RflomicsMAE}(object, selectedResponse = NULL)
}
\arguments{
\item{object}{a MAE object (produced by Flomics).}

\item{response}{a character giving the response variable to access
specifically.}

\item{onlyResults}{default return only the MixOmics or MOFA2 results.
If you want to access all information of the integration,
set onlyResults to FALSE.
In MixOmics case, works only when response is specified.}

\item{results}{The MOFA or mixOmics results to set in the object. 
If null, set to NULL.}

\item{selectedResponse}{a character.
Useful if MixOmics was run on several response variable.
If NULL, all variables are taken into account.}
}
\value{
For getters:
in getMixOmics, if response is NULL,
then all the mixOmics results are returned.
Otherwise, it gives the particular mixOmics result.
For MOFA, returns the untrained object and the trained object as a list.

For setters: always returns a \link{RflomicsMAE} object.

sumMixOmics: A data frame or a list of dataframe
(if selectedResponse is NULL) presenting the summary of mixOmics analyses.
}
\description{
These methods are used to directly access the results of multi-omics 
analyses or their settings, usually stored in the metadata of the 
\link{RflomicsMAE} object. Setters are also available.
}
\examples{

MAEtest <- generateExample(annotation = FALSE, coexp = FALSE)

# Access mixOmics results:
getMixOmics(MAEtest, response = "temperature")
getMixOmicsSettings(MAEtest)
mixOmics::plotIndiv(getMixOmics(MAEtest, response = "imbibition"))

# Access MOFA2 results:
getMOFA(MAEtest)
getMOFASettings(MAEtest)
MOFA2::plot_variance_explained(getMOFA(MAEtest))

}
