% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_07_integration.R
\name{MOFACorNetwork}
\alias{MOFACorNetwork}
\title{Plot correlation network between omics from MOFA2 results}
\usage{
MOFACorNetwork(
  resMOFA,
  factor_choice = 1,
  abs_weight_network = 0.5,
  abs_min_cor_network = 0.5,
  network_layout = "Circle + omics",
  omics_colors = NULL,
  posCol = "red",
  negCol = "blue"
)
}
\arguments{
\item{resMOFA}{a MOFA result}

\item{factor_choice}{chose weight from this factor}

\item{abs_weight_network}{threshold of weight to select entities to print}

\item{abs_min_cor_network}{correlation threshold}

\item{network_layout}{one of spring, circle , circle + omics}

\item{omics_colors}{named list of colors palettes, one for each omics
(can be NULL)}

\item{posCol}{colors of positive edges}

\item{negCol}{colors of negative edges}
}
\value{
A ggplot2 graph
}
\description{
Plot correlation network between omics from MOFA2 results
}
\examples{
MAEtest <- generateExample(coexp = FALSE, annotation = FALSE)
MOFACorNetwork(getMOFA(MAEtest), abs_weight_network = 0.8, 
                                   abs_min_cor_network = 0.8)
}
