% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\docType{methods}
\name{mvQCdata,FlomicsExperiment-method}
\alias{mvQCdata,FlomicsExperiment-method}
\title{multivariate QC data}
\usage{
\S4method{mvQCdata}{FlomicsExperiment}(object, axis = 3)
}
\arguments{
\item{axis}{The number of PCA axis}

\item{FlomicsExperiment}{an object of Class FlomicsExperiment}
}
\description{
multivariate QC data
}
