% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\name{mvQCdesign,MultiAssayExperiment-method}
\alias{mvQCdesign,MultiAssayExperiment-method}
\title{Multivariate Quality Check}
\usage{
\S4method{mvQCdesign}{MultiAssayExperiment}(object, data, PCA = c("raw", "norm"), axis = 5, pngFile = NULL)
}
\arguments{
\item{data}{data name}

\item{PCA}{pca axis to plot}

\item{axis}{The number of PCA axis}

\item{pngFile}{plot file name to save}

\item{MultiAssayExperiment}{an object of Class MultiAssayExperiment}
}
\description{
Multivariate Quality Check
}
