% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Class.R
\docType{class}
\name{Normalization-class}
\alias{Normalization-class}
\alias{.Normalization}
\title{Normalization class}
\value{
An instance of the Normalization class
}
\description{
Normalization class
}
\section{Slots}{

\describe{
\item{\code{method}}{A vector giving the name of the normalization method}

\item{\code{Norm.factors}}{A named vector giving for each samples, its factor of normalization}
}}

