% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\name{Normalization-Constructor}
\alias{Normalization-Constructor}
\alias{Normalization}
\title{[\code{\link{Normalization-class}}] Class constructor}
\usage{
Normalization(ExperimentType)
}
\arguments{
\item{ExperimentType}{The type of experiment (either 'RNAseq' or 'Proteomic' or 'Metabolomic')}
}
\value{
An object of class [\code{\link{Normalization-class}}]
}
\description{

}
