% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_04_diff_analysis.R
\name{plotBoxplotDE,RflomicsSE-method}
\alias{plotBoxplotDE,RflomicsSE-method}
\alias{plotBoxplotDE}
\alias{plotBoxplotDE,RflomicsMAE-method}
\title{plotBoxplotDE}
\usage{
\S4method{plotBoxplotDE}{RflomicsSE}(object, features = NULL, groupColor = "groups", raw = FALSE)

\S4method{plotBoxplotDE}{RflomicsMAE}(
  object,
  SE.name,
  features = NULL,
  groupColor = "groups",
  raw = FALSE
)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE}}

\item{features}{variable name (gene/protein/metabolite name)}

\item{groupColor}{default to groups, indicate a variable in the design to
color the boxplots accordingly.}

\item{raw}{Boolean. Plot the raw data or the transformed ones (TRUE)}

\item{SE.name}{the name of the \link{RflomicsSE} object to fetch if the object is of class \link{RflomicsMAE}}
}
\value{
a ggplot graph
}
\description{
plotBoxplotDE

plotBoxplotDE
}
\examples{

# Set the data path
datPath <- paste0(system.file(package = "RFLOMICS"), "/ExamplesFiles/ecoseed/")

# Load the experimental design
ExpDesign <- readExpDesign(file = paste0(datPath, "condition.txt"))

# Set factor name, levels, level of reference and type
facRef <- data.frame(factorName = c("Repeat", "temperature" , "imbibition"), 
                     factorRef = c("rep1", "Low", "DS"), 
                     factorType = c("batch",  "Bio", "Bio"), 
                     factorLevels = c("rep1,rep2,rep3",
                                      "Low,Medium,Elevated",
                                      "DS,EI,LI"))

# Load the omics data
omicsData <- list( 
readOmicsData(file = paste0(datPath, "proteome_ecoseed.txt")))
 
# Instantiate an object of class RflomicsMAE 
MAE <- createRflomicsMAE(projectName = "Tests", 
                         omicsData   = omicsData,
                         omicsNames  = c("protetest"),
                         omicsTypes  = c("proteomics"),
                         ExpDesign   = ExpDesign, factorRef   = facRef)
names(MAE) <- c("protetest")

# Set the statistical model and contrasts to test
formulae <- generateModelFormulae(MAE)
MAE <- setModelFormula(MAE, formulae[[1]])  

# Get the contrasts List and choose the first 3 contrasts of type averaged
contrastList <- getPossibleContrasts(MAE, formula = formulae[[1]], 
                                     typeContrast = "averaged", 
                                     returnTable = TRUE)[c(1, 2, 3),]

# Run the data preprocessing and perform the differential analysis
 MAE |>  runTransformData(SE.name = "protetest",  transformMethod = "log2") |>
 runNormalization(SE.name = "protetest", normMethod = "median")    |>
 runDiffAnalysis(SE.name = "protetest", method = "limmalmFit", contrastList = contrastList)
 
 # plot the boxplot
 plotBoxplotDE(experiments(MAE)[[1]], features = "AT1G15670", groupColor = "temperature")
 plotBoxplotDE(experiments(MAE)[[1]], features = "AT1G15670", groupColor = "imbibition")
}
