% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_06_annot_enrichment.R
\name{plotClusterProfiler,RflomicsSE-method}
\alias{plotClusterProfiler,RflomicsSE-method}
\alias{plotClusterProfiler}
\title{plotClusterProfiler}
\usage{
\S4method{plotClusterProfiler}{RflomicsSE}(
  object,
  contrastName,
  database,
  domain = "no-domain",
  from = "DiffExp",
  plotType = "dotplot",
  showCategory = 15,
  searchExpr = "",
  nodeLabel = "all",
  p.adj.cutoff = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE}.
It is expected the SE object is produced by rflomics previous analyses,
as it relies on their results.}

\item{contrastName}{the name of the contrast to consider.
For Co expression analysis, it is expected to be one of
"cluster.1", "cluster.2", etc.}

\item{database}{the database (GO, KEGG or custom)}

\item{domain}{if the database is GO, expect one of BP, MF or CC.
Default is NULL.}

\item{from}{what type of analysis to consider?
One of 'DiffExp' or 'CoExp'}

\item{plotType}{type of plot. Define the function used inside.
One of dotplot, heatplot or cnetplot.}

\item{showCategory}{max number of terms to show.}

\item{searchExpr}{expression to search in the showCategory terms.}

\item{nodeLabel}{same as in enrichplot::cnetplot function, defines}

\item{p.adj.cutoff}{pvalueCutoff to define the enrichment threshold.}

\item{...}{additionnal parameters for cnetplot, heatplot or
enrichplot functions.}
}
\value{
A ggplot object.
}
\description{
Plot a dotplot, a cnetplot or an heatplot, using enrichplot
package. It is a wrapper method destined for the RflomicsSE class.
}
\examples{
# Generate RflomicsMAE for example
MAEtest <- generateExample(annotation = FALSE, integration = FALSE)
# Run KEGG annotation on differential analysis results
MAEtest[["protetest"]] <- runAnnotationEnrichment(MAEtest[["protetest"]],
                        list_args = list(organism = "ath",
                                         keyType = "kegg",
                                         pvalueCutoff = 0.05),
                        from = "DiffExp", database = "KEGG")

# From differential analysis proteins lists:
plotClusterProfiler(MAEtest[["protetest"]],
        contrastName = "(temperatureElevated - temperatureMedium) in mean",
        database = "KEGG", from = "DiffExp",
        plotType = "heatplot", p.adj.cutoff = 0.05,
        domain = "no-domain")

}
