% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_05_coexp_analysis.R
\name{plotCoExpressionProfile,RflomicsSE-method}
\alias{plotCoExpressionProfile,RflomicsSE-method}
\alias{plotCoExpressionProfile}
\alias{plotCoExpressionProfile,RflomicsMAE-method}
\title{plotCoExpressionProfile}
\usage{
\S4method{plotCoExpressionProfile}{RflomicsSE}(
  object,
  cluster = 1,
  condition = "groups",
  features = NULL
)

\S4method{plotCoExpressionProfile}{RflomicsMAE}(
  object,
  SE.name,
  cluster = 1,
  condition = "groups",
  features = NULL
)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE}}

\item{cluster}{cluster number}

\item{condition}{Default is group.}

\item{features}{Default is NULL.}

\item{SE.name}{the name of the data to fetch in the object if the object 
is a RflomicsMAE}
}
\value{
a ggplot graph
}
\description{
plotCoExpressionProfile

plotCoExpressionProfile
}
\examples{

# Set the data path
datPath <- paste0(system.file(package = "RFLOMICS"), "/ExamplesFiles/ecoseed/")

# Load the experimental design
ExpDesign <- readExpDesign(file = paste0(datPath, "condition.txt"))

# Set factor name, levels, level of reference and type
facRef <- data.frame(factorName   = c("Repeat", "temperature" , "imbibition"), 
                     factorRef = c("rep1", "Low", "DS"), 
                     factorType = c("batch",  "Bio", "Bio"), 
                     factorLevels = c("rep1,rep2,rep3","Low,Medium,Elevated","DS,EI,LI"))

# Load the omics data
omicsData <- list( 
readOmicsData(file = paste0(datPath, "proteome_ecoseed.txt")))
 
# Instantiate an object of class RflomicsMAE 
MAE <- createRflomicsMAE( projectName = "Tests", omicsData   = omicsData,
                         omicsNames  = c("protetest"), omicsTypes  = c("proteomics"),
                         ExpDesign   = ExpDesign, factorRef   = facRef)
names(MAE) <- c("protetest")

# Set the statistical model and contrasts to test
formulae <- generateModelFormulae(MAE)
MAE <- setModelFormula(MAE, formulae[[1]])  

# Get the contrasts List and choose the first 3 contrasts of type averaged
contrastList <- getPossibleContrasts(MAE, formula = formulae[[1]],
                                     typeContrast = "averaged", 
                                     returnTable = TRUE)[c(1, 2, 3),]

# Run the data preprocessing and perform the differential analysis
 MAE <- MAE |>  runTransformData(SE.name = "protetest",  transformMethod = "log2") |>
 runNormalization(SE.name = "protetest", normMethod = "median")    |> 
 runDiffAnalysis(SE.name = "protetest",  method = "limmalmFit", contrastList = contrastList) |> 
 runCoExpression(SE.name = "protetest", K = 2:10, replicates = 5, 
                 contrastNames = contrastList$tag, merge = "union") 
 
 plotCoExpressionProfile(MAE[["protetest"]], cluster = 2)
 
}
