% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_01_load_data.R
\name{plotConditionsOverview,RflomicsMAE-method}
\alias{plotConditionsOverview,RflomicsMAE-method}
\alias{plotConditionsOverview}
\title{plotConditionsOverview}
\usage{
\S4method{plotConditionsOverview}{RflomicsMAE}(object, omicNames = NULL)
}
\arguments{
\item{object}{An object of class \link{RflomicsMAE-class}}

\item{omicNames}{a vector with list of dataset names}
}
\value{
a gg plot object
\itemize{
 \item{"plot:"}{ plot of count data.frame.}
 }
}
\description{
A complete design and at least one biological and one batch factors are 
required for using RFLOMICS workflow.
}
