% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_05_coexp_analysis.R
\name{plotCoseqContrasts,RflomicsSE-method}
\alias{plotCoseqContrasts,RflomicsSE-method}
\alias{plotCoseqContrasts}
\alias{plotCoseqContrasts,RflomicsMAE-method}
\title{plotCoseqContrasts
This function describes the composition of clusters according to the 
contrast to which the gene belongs}
\usage{
\S4method{plotCoseqContrasts}{RflomicsSE}(object)

\S4method{plotCoseqContrasts}{RflomicsMAE}(object, SE.name)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE}}

\item{SE.name}{name of the RflomicsSE to access}
}
\value{
a ggplot graph
}
\description{
plotCoseqContrasts
This function describes the composition of clusters according to the 
contrast to which the gene belongs

plotCoseqContrasts
}
\examples{

# Set the data path
datPath <- paste0(system.file(package = "RFLOMICS"), "/ExamplesFiles/ecoseed/")

# Load the experimental design
ExpDesign <- readExpDesign(file = paste0(datPath, "condition.txt"))

# Set factor name, levels, level of reference and type
facRef <- data.frame(factorName   = c("Repeat", "temperature" , "imbibition"), 
                     factorRef = c("rep1", "Low", "DS"), 
                     factorType = c("batch",  "Bio", "Bio"),
                     factorLevels = c("rep1,rep2,rep3","Low,Medium,Elevated","DS,EI,LI"))

# Load the omics data
omicsData <- list( 
readOmicsData(file = paste0(datPath, "proteome_ecoseed.txt")))
 
# Instantiate an object of class RflomicsMAE 
MAE <- createRflomicsMAE( projectName = "Tests", omicsData   = omicsData,
                         omicsNames  = c("protetest"), omicsTypes  = c("proteomics"),
                         ExpDesign   = ExpDesign, factorRef   = facRef)
names(MAE) <- c("protetest")

# Set the statistical model and contrasts to test
formulae <- generateModelFormulae(MAE)
MAE <- setModelFormula(MAE, formulae[[1]])  

# Get the contrasts List and choose the first 3 contrasts of type averaged
contrastList <- getPossibleContrasts(MAE, formula = formulae[[1]],
                                     typeContrast = "averaged",
                                     returnTable = TRUE)[c(1, 2, 3),]

# Run the data preprocessing and perform the differential analysis
 MAE <- MAE |>  runTransformData(SE.name = "protetest",  transformMethod = "log2") |>
 runNormalization(SE.name = "protetest", normMethod = "median")    |> 
 runDiffAnalysis(SE.name = "protetest",  method = "limmalmFit", contrastList = contrastList) |> 
 runCoExpression(SE.name = "protetest", K = 2:10, replicates = 5, 
                 contrastNames = contrastList$tag, merge = "union") 
 
 plotCoseqContrasts(MAE[["protetest"]])
}
