% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_03_data_explor.R
\name{plotDataDistribution,RflomicsSE-method}
\alias{plotDataDistribution,RflomicsSE-method}
\alias{plotDataDistribution}
\alias{plotDataDistribution,RflomicsMAE-method}
\title{plotDataDistribution}
\usage{
\S4method{plotDataDistribution}{RflomicsSE}(object, plot = "boxplot", raw = FALSE)

\S4method{plotDataDistribution}{RflomicsMAE}(object, SE.name, plot = "boxplot", raw = FALSE)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE}}

\item{plot}{plot type ("boxplot" or "density")}

\item{raw}{boolean. Plot the raw data or the transformed ones (raw = FALSE)}

\item{SE.name}{the name of the data the normalization have to be applied to.}
}
\value{
A ggplot object
}
\description{
plotDataDistribution

plotDataDistribution
}
\examples{

# Set the data path
datPath <- paste0(system.file(package = "RFLOMICS"), "/ExamplesFiles/ecoseed/")

# Load the experimental design
ExpDesign <- readExpDesign(file = paste0(datPath, "condition.txt"))

# Set factor name, levels, level of reference and type
facRef <- data.frame(factorName   = c("Repeat", "temperature" , "imbibition"), 
                     factorRef = c("rep1", "Low", "DS"), 
                     factorType = c("batch",  "Bio", "Bio"), 
                     factorLevels = c("rep1,rep2,rep3","Low,Medium,Elevated","DS,EI,LI"))

# Load the omics data
omicsData <- list( 
readOmicsData(file = paste0( datPath, "transcriptome_ecoseed.txt")), 
readOmicsData(file = paste0(datPath, "proteome_ecoseed.txt")))
 
# Instantiate an object of class RflomicsMAE 
MAE <- createRflomicsMAE(projectName = "Tests", omicsData   = omicsData,
                         omicsNames  = c("RNAtest", "protetest"),
                         omicsTypes  = c("RNAseq","proteomics"),
                         ExpDesign   = ExpDesign, factorRef   = facRef)
names(MAE) <- c("RNAtest", "protetest")

# Set the statistical model and contrasts to test
formulae <- generateModelFormulae(MAE)
MAE <- setModelFormula(MAE, formulae[[1]])  

# Get the contrasts List and choose the first 3 contrasts of type averaged
contrastList <- getPossibleContrasts(MAE,
                                     formula = formulae[[1]], 
                                     typeContrast = "averaged", 
                                     returnTable = TRUE)[c(1, 2, 3),]

# Run the data preprocessing and perform the differential analysis
 MAE |>  runTransformData(SE.name = "protetest",  transformMethod = "log2") |>
 filterLowAbundance(SE.name = "RNAtest")                           |>    
 runNormalization(SE.name = "RNAtest",   normMethod = "TMM")       |>
 runNormalization(SE.name = "protetest", normMethod = "median")
 
 # plotLibrarySize
 plotDataDistribution(experiments(MAE)[[1]], raw=FALSE)
 plotDataDistribution(experiments(MAE)[[1]], raw=TRUE)
 
}
