% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_01_load_data.R
\name{plotDataOverview,RflomicsMAE-method}
\alias{plotDataOverview,RflomicsMAE-method}
\alias{plotDataOverview}
\title{Dataset overview plot}
\usage{
\S4method{plotDataOverview}{RflomicsMAE}(object, omicNames = NULL, realSize = FALSE)
}
\arguments{
\item{object}{An object of class \link{RflomicsMAE-class}}

\item{omicNames}{a vector with dataset names}

\item{realSize}{booleen value}
}
\value{
plot
}
\description{
This function plot an overview of the loaded datasets displaying per sample 
(n=number of entities (genes/metabolites/proteins); k=number of samples)
}
