% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_04_diff_analysis.R
\name{plotDiffAnalysis,RflomicsSE-method}
\alias{plotDiffAnalysis,RflomicsSE-method}
\alias{plotDiffAnalysis}
\alias{plotDiffAnalysis,RflomicsMAE-method}
\title{plotDiffAnalysis}
\usage{
\S4method{plotDiffAnalysis}{RflomicsSE}(
  object,
  contrastName,
  typeofplots = c("MA.plot", "volcano", "histogram")
)

\S4method{plotDiffAnalysis}{RflomicsMAE}(
  object,
  SE.name,
  contrastName,
  typeofplots = c("MA.plot", "volcano", "histogram")
)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE}}

\item{contrastName}{The contrastName for which the plots has to be drawn}

\item{typeofplots}{The plots you want to return. Default is all possible 
plots: MA plot, Volcano plot and non adjusted pvalues histogram.}

\item{SE.name}{the name of the data to fetch in the object if the object 
is a RflomicsMAE}
}
\value{
plot
}
\description{
This is an interface method which draw a MAplot, a volcano plot and the 
pvalues distribution from the results of a differential analysis
performed on omic datasets stored in an object of class \link{RflomicsSE}
}
