% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotDistr}
\alias{plotDistr}
\title{plotDistr}
\usage{
plotDistr(abundances, dataName, pngFile = NULL)
}
\arguments{
\item{abundances}{matrix or dataframe of feature/gene abundances/counts}

\item{dataName}{name of dataset}

\item{pngFile}{png file name}
}
\value{
plot
}
\description{
plotDistr
}
\examples{
plotDistr(assay(MAE), "dataset1", "tmp/countDist.png")
}
