% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_06_annot_enrichment.R
\name{plotEnrichComp,RflomicsSE-method}
\alias{plotEnrichComp,RflomicsSE-method}
\alias{plotEnrichComp}
\title{plotEnrichComp}
\usage{
\S4method{plotEnrichComp}{RflomicsSE}(
  object,
  from = "DiffExp",
  database = NULL,
  domain = "no-domain",
  matrixType = "FC",
  nClust = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE}.
It is expected the SE object is produced by rflomics previous analyses,
as it relies on their results.}

\item{from}{indicates if the enrichment results are taken from differential
analysis results (DiffExpAnal) or from the co-expression analysis results
 (CoExpAnal)}

\item{database}{is it a custom annotation, GO or KEGG annotations}

\item{domain}{domain from the database (eg GO has three domains,
BP, CC and MF)}

\item{matrixType}{Heatmap matrix to plot, one of GeneRatio, p.adjust or
presence.}

\item{nClust}{number of separate cluster to plot on the heatmap, based o
n the clustering.}

\item{...}{more arguments for ComplexHeatmap::Heatmap.}
}
\value{
A complexHeatmap object.
}
\description{
Plot an heatmap of all the enriched term found for a given
database and a given source (differential analysis or coexpression clusters).
Allow for the comparison of several enrichment results.
}
\examples{
# Generate RflomicsMAE for example
MAEtest <- generateExample(annotation = FALSE, integration = FALSE)
# Run KEGG annotation on differential analysis results
MAEtest[["protetest"]] <- runAnnotationEnrichment(MAEtest[["protetest"]],
                        list_args = list(organism = "ath",
                                         keyType = "kegg",
                                         pvalueCutoff = 0.1),
                        from = "DiffExp", database = "KEGG")

plotEnrichComp(MAEtest[["protetest"]], from = "DiffExp",
               database = "KEGG", matrixType = "FC")

}
