% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_01_load_data.R
\name{plotExpDesignCompleteness,RflomicsSE-method}
\alias{plotExpDesignCompleteness,RflomicsSE-method}
\alias{plotExpDesignCompleteness}
\alias{plotExpDesignCompleteness,RflomicsMAE-method}
\title{plotExpDesignCompleteness}
\usage{
\S4method{plotExpDesignCompleteness}{RflomicsSE}(object, sampleList = NULL)

\S4method{plotExpDesignCompleteness}{RflomicsMAE}(object, omicName, sampleList = NULL)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE-class}}

\item{sampleList}{list of samples to take into account}

\item{omicName}{a character string with the name of the dataset}
}
\value{
plot
}
\description{
This method checks that experimental design constraints are satisfied and plot 
a summary of the design.
A complete design (all combinations of factor modalities with at least 2 replicates for each have to be present) 
with at least one biological and one batch factors are required to use the RFLOMICS workflow.
}
