% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_04_diff_analysis.R
\name{plotHeatmapDesign,RflomicsSE-method}
\alias{plotHeatmapDesign,RflomicsSE-method}
\alias{plotHeatmapDesign}
\alias{plotHeatmapDesign,RflomicsMAE-method}
\title{plotHeatmapDesign}
\usage{
\S4method{plotHeatmapDesign}{RflomicsSE}(
  object,
  contrastName,
  splitFactor = "none",
  title = "",
  annotNames = NULL,
  modalities = NULL,
  drawArgs = list(),
  heatmapArgs = list()
)

\S4method{plotHeatmapDesign}{RflomicsMAE}(
  object,
  SE.name,
  contrastName,
  splitFactor = "none",
  title = "",
  annotNames = NULL,
  modalities = NULL,
  drawArgs = list(),
  heatmapArgs = list()
)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE}}

\item{contrastName}{The contrastName for which the MAplot has to be drawn}

\item{splitFactor}{characters. Default to none. Name of a feature in the 
design matrix, splits the samples on the heatmap according to its modalities.}

\item{title}{characters. Title of the heatmap.}

\item{annotNames}{vector. Names of the annotations to keep in the Heatmap. 
Default takes all available information.}

\item{modalities}{named list of vectors of modalities to subset and print 
on the heatmap.}

\item{drawArgs, heatmapArgs}{named lists. Any additional parameter passed 
to ComplexHeatmap::Heatmap or ComplexHeatmap::draw}

\item{SE.name}{the name of the data to fetch in the object if the object 
is a RflomicsMAE}
}
\value{
plot
}
\description{
This mehod draws a heatmap from the results of a 
differential analysis performed on omic datasets stored in an object of 
class \link{RflomicsSE}
}
\examples{

# Set the data path
datPath <- paste0(system.file(package = "RFLOMICS"), "/ExamplesFiles/ecoseed/")

# Load the experimental design
ExpDesign <- readExpDesign(file = paste0(datPath, "condition.txt"))

# Set factor name, levels, level of reference and type
facRef <- data.frame(factorName   = c("Repeat", "temperature" , "imbibition"),
                     factorRef = c("rep1", "Low", "DS"), 
                     factorType = c("batch",  "Bio", "Bio"), 
                     factorLevels = c("rep1,rep2,rep3","Low,Medium,Elevated","DS,EI,LI"))

# Load the omics data
omicsData <- list( 
readOmicsData(file = paste0(datPath, "proteome_ecoseed.txt")))
 
# Instantiate an object of class RflomicsMAE 
MAE <- createRflomicsMAE(projectName = "Tests", omicsData   = omicsData,
                         omicsNames  = c("protetest"),omicsTypes  = c("proteomics"),
                         ExpDesign   = ExpDesign, factorRef   = facRef)
names(MAE) <- c("protetest")

# Set the statistical model and contrasts to test
formulae <- generateModelFormulae(MAE)
MAE <- setModelFormula(MAE, formulae[[1]])  

# Get the contrasts List and choose the first 3 contrasts of type averaged
contrastList <- getPossibleContrasts(MAE, formula = formulae[[1]], 
                                     typeContrast = "averaged", 
                                     returnTable = TRUE)[c(1, 2, 3),]

# Run the data preprocessing and perform the differential analysis
MAE <- MAE |>  runTransformData(SE.name = "protetest",  transformMethod = "log2") |>
 runNormalization(SE.name = "protetest", normMethod = "median")    |>
 runDiffAnalysis(SE.name = "protetest", method = "limmalmFit", 
                  contrastList = contrastList)
 
 # plot the heatmap
 plotHeatmapDesign(experiments(MAE)[[1]], 
                   getSelectedContrasts(experiments(MAE)[[1]])$contrastName[1])
}
