% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_06_annot_enrichment.R
\name{plotKEGG,RflomicsSE-method}
\alias{plotKEGG,RflomicsSE-method}
\alias{plotKEGG}
\title{plotKEGG}
\usage{
\S4method{plotKEGG}{RflomicsSE}(
  object,
  contrastName,
  pathway_id = NULL,
  species = "ath",
  gene_idtype = "kegg",
  from = "DiffExp",
  ...
)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE}.
It is expected the SE object is produced by rflomics previous analyses,
as it relies on their results.}

\item{contrastName}{the name of the contrast to consider.
For Co expression analysis,
it is expected to be one of "cluster.1", "cluster.2", etc.}

\item{pathway_id}{the KEGG id pathway to plot.}

\item{species}{Kegg code (eg hsa or ath)}

\item{gene_idtype}{idtype (kegg, uniprot,...)}

\item{from}{differential analysis (diffExp) or coexpression analysis (coexp)
Used to link contrastName to the right metadata.}

\item{...}{Not in use at the moment}
}
\value{
Only displays the KEGG pathway, it does not return any object.
}
\description{
Plot the KEGG pathway using the pathview package. It overlays
the result of the differential analysis or the clusters entity on the
image.
}
\examples{
# Generate RflomicsMAE for example
 datPath <- paste0(system.file(package = "RFLOMICS"),
                   "/ExamplesFiles/ecoseed/")

ExpDesign <- readExpDesign(file = paste0(datPath, "condition.txt"))
facRef <- data.frame(
                factorName   = c("Repeat", "temperature" , "imbibition"),
                factorRef    = c("rep1",   "Low",          "DS"),
                factorType   = c("batch",  "Bio",          "Bio"),
                factorLevels = c("rep1,rep2,rep3",
                                 "Low,Medium,Elevated",
                                 "DS,EI,LI"))
                       
omicsData <- list(
   readOmicsData(file = paste0(datPath, "transcriptome_ecoseed.txt")),
   readOmicsData(file = paste0(datPath, "metabolome_ecoseed.txt")))
   
MAEtest <- createRflomicsMAE(projectName = "Tests",
                             omicsData   = omicsData,
                             omicsNames  = c("RNAtest",  "protetest"),
                             omicsTypes  = c("RNAseq", "proteomics"),
                             ExpDesign   = ExpDesign,
                             factorRef   = facRef
                             )
names(MAEtest) <- c("RNAtest", "protetest")

# Run KEGG annotation (enrichKEGG)
MAEtest[["protetest"]] <- runAnnotationEnrichment(MAEtest[["protetest"]],
                       list_args = list(organism = "ath",
                       keyType = "kegg",
                       pvalueCutoff = 0.05),
                       from = "CoExp", database = "KEGG")
                       
plotKEGG(MAEtest[["protetest"]], pathway_id = "ath00710", species = "ath",
          contrastName = "cluster.4", from = "Coexp")

}
