% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_00_Common.R
\name{plotOmicsPCA,RflomicsSE-method}
\alias{plotOmicsPCA,RflomicsSE-method}
\alias{plotOmicsPCA}
\alias{plotOmicsPCA,RflomicsMAE-method}
\title{plotOmicsPCA}
\usage{
\S4method{plotOmicsPCA}{RflomicsSE}(
  object,
  raw = c("raw", "norm"),
  axes = c(1, 2),
  groupColor = "groups"
)

\S4method{plotOmicsPCA}{RflomicsMAE}(object, SE.name, raw, axes = c(1, 2), groupColor = "groups")
}
\arguments{
\item{object}{An object of class \link{RflomicsSE-class}}

\item{raw}{This argument indicates whether the scaled PCA has to be 
performed on raw [\sQuote{raw}] or normalized [\sQuote{norm}] data.}

\item{axes}{A vector giving the two axis that have to be drawn for the 
factorial map}

\item{groupColor}{All combination of level's factor}

\item{SE.name}{the name of the data the normalization have to be applied to.}
}
\value{
PCA plot (ggplot2 object)
}
\description{
This function plot the factorial map from a PCA object stored
in a \link{RflomicsSE-class} object. By default, samples are
colored by groups (all combinations of level's factor)
}
