% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\docType{methods}
\name{plotPCAnorm,FlomicsExperiment-method}
\alias{plotPCAnorm,FlomicsExperiment-method}
\title{plotPCAnorm}
\usage{
\S4method{plotPCAnorm}{FlomicsExperiment}(object, data = "norm",
  PCs = c(1, 2), condition = "groups")
}
\arguments{
\item{FlomicsExperiment}{}

\item{color}{color palette}
}
\description{
plotPCAnorm
}
