% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\name{plotPCAnorm,MultiAssayExperiment-method}
\alias{plotPCAnorm,MultiAssayExperiment-method}
\title{plotPCAnorm}
\usage{
\S4method{plotPCAnorm}{MultiAssayExperiment}(object, data, PCA, PCs = c(1, 2), condition = "groups", pngFile)
}
\arguments{
\item{color}{color palette}
}
\value{

}
\description{
plotPCAnorm
}
