% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_01_load_data.R
\name{readExpDesign}
\alias{readExpDesign}
\title{Read Experimental Design}
\usage{
readExpDesign(file)
}
\arguments{
\item{file}{path to experimental design file}
}
\value{
data.frame
}
\description{
Read Experimental Design
}
\examples{
design <- readExpDesign(paste0(system.file(package = "RFLOMICS"), 
                    "/ExamplesFiles/ecoseed/condition.txt"))

}
