% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_01_load_data.R
\name{readOmicsData}
\alias{readOmicsData}
\title{Read omics data}
\usage{
readOmicsData(file)
}
\arguments{
\item{file}{omics data matrix}
}
\value{
data.frame
}
\description{
Read omics data
}
\examples{
readOmicsData(file = paste0(system.file(package = "RFLOMICS"),
              "/ExamplesFiles/ecoseed/transcriptome_ecoseed.txt"))

}
