% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_01_load_data.R
\name{RflomicsMAE-methods}
\alias{RflomicsMAE-methods}
\alias{getProjectName,RflomicsMAE-method}
\alias{getProjectName}
\alias{getDesignMat,RflomicsMAE-method}
\alias{getDesignMat-RflomicsMAE}
\alias{getDatasetNames,RflomicsMAE-method}
\alias{getDatasetNames-RflomicsMAE}
\alias{getOmicsTypes,RflomicsMAE-method}
\alias{getOmicsTypes-RflomicsMAE}
\alias{getFactorNames,RflomicsMAE-method}
\alias{getFactorNames-RflomicsMAE}
\alias{getFactorTypes,RflomicsMAE-method}
\alias{getFactorTypes}
\alias{getBioFactors,RflomicsMAE-method}
\alias{getBioFactors-RflomicsMAE}
\alias{getBatchFactors,RflomicsMAE-method}
\alias{getBatchFactors-RflomicsMAE}
\alias{getMetaFactors,RflomicsMAE-method}
\alias{getMetaFactors-RflomicsMAE}
\alias{getRflomicsSE,RflomicsMAE-method}
\alias{getRflomicsSE}
\alias{getFactorModalities,RflomicsMAE-method}
\alias{getFactorModalities-RflomicsMAE}
\alias{subRflomicsMAE,RflomicsMAE-method}
\alias{subRflomicsMAE}
\title{Access or modify the metadata slot of an object of class RflomicsMAE}
\usage{
\S4method{getProjectName}{RflomicsMAE}(object)

\S4method{getDesignMat}{RflomicsMAE}(object)

\S4method{getDatasetNames}{RflomicsMAE}(object)

\S4method{getOmicsTypes}{RflomicsMAE}(object)

\S4method{getFactorNames}{RflomicsMAE}(object)

\S4method{getFactorTypes}{RflomicsMAE}(object)

\S4method{getBioFactors}{RflomicsMAE}(object)

\S4method{getBatchFactors}{RflomicsMAE}(object)

\S4method{getMetaFactors}{RflomicsMAE}(object)

\S4method{getRflomicsSE}{RflomicsMAE}(object, datasetName = NULL)

\S4method{getFactorModalities}{RflomicsMAE}(object, factorName)

\S4method{subRflomicsMAE}{RflomicsMAE}(object, omicNames = NULL)
}
\arguments{
\item{object}{A \code{RflomicsMAE} object}

\item{datasetName}{name of omic dataset}

\item{factorName}{a name of factor}

\item{omicNames}{a vector with the dataset names}
}
\value{
Accessors: Either a \code{sampleMap}, \code{ExperimentList}, or
\code{DataFrame} object

Setters: A \code{RflomicsMAE}  object

a character string with the project name

a data frame with design matrix

a vector with omic dataset names

names vector

a vector with the names of the design factors

a named vector

a vector with biological factors

a vector with batch factors

a vector with batch factors

An object of class \link{RflomicsSE-class}

a vector

object An object of class \link{RflomicsMAE-class}
}
\description{
A set of accessor and setter generic functions to access and modify 
objects of the slot metadata of a \link{RflomicsMAE} object.
}
\section{Accessors}{

\itemize{
   \item getProjectName: Access to the name of the omic analysis.
   \item getDesignMat: Get the design matrix associated the omic analysis.
   \item getDatasetNames: Names of the loaded omics datasets.
   \item getOmicsTypes: Access to the type of omics datasets: "RNAseq", "Proteomics", "Metabolomics"
   \item getFactorNames: List the experimental factor names.
   \item getFactorTypes: List the experimental factor types ("bio", "batch" or "meta") containing in the design list of the metadata.
   \item getBioFactors: List the biological factors.
   \item getBatchFactors: List the batch factors.
   \item getMetaFactors: List the metadata factors.
   \item getFactorModalities: Access to the factor modalities 
   \item getRflomicsSE: Access to a particular RflomicsSE object 
   \item subRflomics: Extract a subset of a RflomicsMAE object.
   \item getModelFormula: Access to the model formula of the statistical analysis
   \item getSelectedContrasts: List the selected contrasts 
   \item getValidContrasts: List the valid contrasts
   \item getContrastMatrix: Get the contrasts matrix 
   \item getTransSettings: Transformation settings of a given omic dataset 
   \item getFilterSettings: Access to the filtering settings of a given omic dataset 
   \item getFilteredFeatures: Access to the filtered features of a given omic dataset 
   \item getNormSettings: Access to the normalization settings of a given omic dataset 
   \item getCoeffNorm: Access to the normalization coefficients of a given omic dataset 
   \item getDiffSettings: Access to the differential expression analysis settings of a given omic dataset
   \item getCoexpSettings : Access to the co-expression analysis settings of a given omic dataset
   \item getClusterEntities: Access to the cluster entities of a given omic dataset
   \item getCoseqClusters: Access to the co-expression clusters of a given omic dataset
}
}

\section{Setters}{

Setter method values (i.e., '\code{function(x) <- value}'):
\itemize{
    \item setModelFormula: Set the model formula stored in \code{metadata} slot.
    \item setSelectedContrasts: Set the selected contrasts stored in \code{metadata} slot.
    \item setValidContrasts: Set the valid contrasts stored in \code{metadata} slot.
    \item setTrans: Set the transformation settings stored in \code{metadata} slot.
    \item setNorm: Set the normalization settings stored in \code{metadata} slot.
}
}

