% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_01_load_data.R
\name{RflomicsMAE}
\alias{RflomicsMAE}
\title{RflomicsMAE}
\usage{
RflomicsMAE(
  experiments = ExperimentList(),
  colData = S4Vectors::DataFrame(),
  sampleMap = S4Vectors::DataFrame(assay = factor(), primary = character(), colname =
    character()),
  omicList = list(),
  projectName = NULL,
  design = list(),
  IntegrationAnalysis = list()
)
}
\arguments{
\item{name}{description}
}
\value{
An object of class \link{RflomicsMAE-class}
}
\description{
\link{RflomicsMAE} constructor.
}
\examples{

rflomicsMAE <- RflomicsMAE()
}
\seealso{
\link{MultiAssayExperiment}
}
