% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_01_load_data.R
\name{RflomicsSE-methods}
\alias{RflomicsSE-methods}
\alias{getDesignMat,RflomicsSE-method}
\alias{getDesignMat-RflomicsSE}
\alias{getDatasetNames,RflomicsSE-method}
\alias{getDatasetNames-RflomicsSE}
\alias{getOmicsTypes,RflomicsSE-method}
\alias{getOmicsTypes-RflomicsSE}
\alias{getFactorNames,RflomicsSE-method}
\alias{getFactorNames-RflomicsSE}
\alias{getFactorTypes,RflomicsSE-method}
\alias{getFactorTypes-RflomicsSE}
\alias{getBioFactors,RflomicsSE-method}
\alias{getBioFactors-RflomicsSE}
\alias{getBatchFactors,RflomicsSE-method}
\alias{getBatchFactors-RflomicsSE}
\alias{getMetaFactors,RflomicsSE-method}
\alias{getMetaFactors-RflomicsSE}
\alias{getFactorModalities,RflomicsSE-method}
\alias{getFactorModalities-RflomicsSE}
\title{Accessing and modifying information in RflomicsSE}
\usage{
\S4method{getDesignMat}{RflomicsSE}(object)

\S4method{getDatasetNames}{RflomicsSE}(object)

\S4method{getOmicsTypes}{RflomicsSE}(object)

\S4method{getFactorNames}{RflomicsSE}(object)

\S4method{getFactorTypes}{RflomicsSE}(object)

\S4method{getBioFactors}{RflomicsSE}(object)

\S4method{getBatchFactors}{RflomicsSE}(object)

\S4method{getMetaFactors}{RflomicsSE}(object)

\S4method{getFactorModalities}{RflomicsSE}(object, factorName)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE-class}}

\item{factorName}{a name of factor}
}
\value{
Accessors: an object stored in the metadata slot of the \code{RflomicsSE} object

Setters: A \code{RflomicsSE} object

a data frame with design matrix

a vector with the omic dataset name

named vector

a vector with names of design factors

a named vector

a vector with biological factors

a vector with batch factors

a vector with batch factors

a vector
}
\description{
A set of accessor and setter generic functions to modify 
objects from the slot metadata of a \code{RflomicsSE} object. 
The slot metadata contains: 

\itemize{
  \item omicType: the type of omics dataset
  \item design: a list of design
  \item DataProcessing: a list of data processing settings and results 
  \item PCAlist: a list of PCA settings and results
  \item DiffExpAnal: a list of DiffExpAnal settings and results
  \item CoExpAnal: a list of CoExpAnal settings and results
  \item DiffExpEnrichAnal: a list of DiffExpEnrichAnal settings and results
  \item CoExpEnrichAnal: a list of CoExpEnrichAnal settings and results
 }
}
\section{Accessors}{

\itemize{
   \item getDesignMat: Get the design matrix associated the omic analysis.
   \item getDatasetNames: Names of the loaded omics datasets.
   \item getOmicsTypes: Access to the type of omics datasets: "RNAseq", "Proteomics", "Metabolomics"
   \item getFactorNames: List the experimental factor names.
   \item getFactorTypes: List the experimental factor types ("bio", "batch" or "meta") containing in the design list of the metadata.
   \item getBioFactors: List the biological factors.
   \item getBatchFactors: List the batch factors.
   \item getMetaFactors: List the metadata factors.
   \item getFactorModalities: Access to the factor modalities 
   \item getRflomicsSE: Access to a particular RflomicsSE object 
   \item subRflomics: Extract a subset of a RflomicsMAE object.
   \item getModelFormula: Access to the model formula of the statistical analysis
   \item getSelectedContrasts: List the selected contrasts 
   \item getValidContrasts: List the valid contrasts
   \item getContrastMatrix: Get the contrasts matrix 
   \item getTransSettings: Transformation settings of a given omic dataset 
   \item getFilterSettings: Access to the filtering settings of a given omic dataset 
   \item getFilteredFeatures: Access to the filtered features of a given omic dataset 
   \item getNormSettings: Access to the normalization settings of a given omic dataset 
   \item getCoeffNorm: Access to the normalization coefficients of a given omic dataset 
   \item getDiffSettings: Access to the differential expression analysis settings of a given omic dataset
   \item getCoexpSettings : Access to the co-expression analysis settings of a given omic dataset
   \item getClusterEntities: Access to the cluster entities of a given omic dataset
   \item getCoseqClusters: Access to the co-expression clusters of a given omic dataset
}
}

\section{Setters}{

Setter method values (i.e., '\code{function(x) <- value}'):
\itemize{
    \item setValid
}
}

