% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\name{RunDiffAnalysis,MultiAssayExperiment-method}
\alias{RunDiffAnalysis,MultiAssayExperiment-method}
\title{RunDiffAnalysis}
\usage{
\S4method{RunDiffAnalysis}{MultiAssayExperiment}(object, data, FDR, DiffAnalysisMethod, clustermq)
}
\arguments{
\item{data}{omic data type}

\item{clustermq}{A boolean indicating if the constrasts have to be computed in local or in a distant machine}

\item{DiffMethod}{Differential analysis method}
}
\value{
MultiAssayExperiment
}
\description{
RunDiffAnalysis
}
\examples{

}
