% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\docType{methods}
\name{RunNormalization,FlomicsExperiment-method}
\alias{RunNormalization,FlomicsExperiment-method}
\title{RunNormalization}
\usage{
\S4method{RunNormalization}{FlomicsExperiment}(object,
  NormMethod = object@Normalization@Method)
}
\arguments{
\item{FlomicsExperiment}{}
}
\value{
FlomicsExperiment
}
\description{
RunNormalization
}
\examples{

}
