% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\name{RunNormalization,MultiAssayExperiment-method}
\alias{RunNormalization,MultiAssayExperiment-method}
\title{RunNormalization}
\usage{
\S4method{RunNormalization}{MultiAssayExperiment}(object, data, NormMethod)
}
\arguments{
\item{data}{omic data type}

\item{NormMethod}{normalisation methode}
}
\value{
MultiAssayExperiment
}
\description{
RunNormalization
}
\examples{

}
