% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_03_data_explor.R
\name{runNormalization,RflomicsSE-method}
\alias{runNormalization,RflomicsSE-method}
\alias{runNormalization}
\alias{runNormalization,RflomicsMAE-method}
\title{runNormalization}
\usage{
\S4method{runNormalization}{RflomicsSE}(object, normMethod = NULL, modifyAssay = FALSE)

\S4method{runNormalization}{RflomicsMAE}(object, SE.name, normMethod = NULL, modifyAssay = FALSE)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE}}

\item{normMethod}{Normalization method}

\item{modifyAssay}{Does the normalization have to be applied or just stored for later? Recommended it stays FALSE.}

\item{SE.name}{the name of the data the normalization have to be applied to.}
}
\value{
An object of class \link{RflomicsSE}
The applied normalization method and computed scaling factors (by samples) are stored as a named list
("normalization") of two elements (respectively "method" and "coefNorm") in the metadata slot of a
given data set, stored itself in the ExperimentList slot of a \link{RflomicsSE} object.
}
\description{
This function applied a normalization method on an omic data sets stored in an object of
class \link{RflomicsSE} or \link{RflomicsMAE}. In this case, the normalization method is applied to the data 
set and 

\itemize{
\item{For RNAseq data:}{the TMM function of edgeR is proposed by default, see the ref}
\item{For Proteomic data:}{}
\item{For Metabolomic data:}{}
}
}
\examples{

# Set the data path
datPath <- paste0(system.file(package = "RFLOMICS"), "/ExamplesFiles/ecoseed/")

# Load the experimental design
ExpDesign <- readExpDesign(file = paste0(datPath, "condition.txt"))

# Set factor name, levels, level of reference and type
facRef <- data.frame(factorName   = c("Repeat", "temperature" , "imbibition"), 
                     factorRef = c("rep1", "Low", "DS"), 
                     factorType = c("batch",  "Bio", "Bio"), 
                     factorLevels = c("rep1,rep2,rep3","Low,Medium,Elevated","DS,EI,LI"))

# Load the omics data
omicsData <- list( 
readOmicsData(file = paste0( datPath, "transcriptome_ecoseed.txt")))
 
# Instantiate an object of class RflomicsMAE 
MAE <- createRflomicsMAE( projectName = "Tests", omicsData   = omicsData,
                         omicsNames  = c("RNAtest"), omicsTypes  = c("RNAseq"),
                         ExpDesign   = ExpDesign, factorRef   = facRef)
names(MAE) <- c("RNAtest")

# Set the statistical model and contrasts to test
formulae <- generateModelFormulae(MAE)
MAE <- setModelFormula(MAE, formulae[[1]])  

# Get the contrasts List and choose the first 3 contrasts of type averaged
contrastList <- getPossibleContrasts(MAE, formula = formulae[[1]], 
                                     typeContrast = "averaged", 
                                     returnTable = TRUE)[c(1),]

# Run the data preprocessing and perform the differential analysis
MAE |>  filterLowAbundance(SE.name = "RNAtest") |>
runNormalization(SE.name = "RNAtest", normMethod = "TMM") 

# Get the normalization method and the scaling factors
getCoeffNorm(experiments(MAE)[[1]])

# Get the normalization settings
getNormSettings(experiments(MAE)[[1]])
}
\references{
Lambert, I., Paysant-Le Roux, C., Colella, S. et al. 
DiCoExpress: a tool to process multifactorial
RNAseq experiments from quality controls to co-expression analysis through 
differential analysis based on contrasts inside GLM models. 
Plant Methods 16, 68 (2020).
}
