% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_00_Common.R
\name{runOmicsPCA,RflomicsSE-method}
\alias{runOmicsPCA,RflomicsSE-method}
\alias{runOmicsPCA}
\alias{runOmicsPCA,RflomicsMAE-method}
\title{runOmicsPCA}
\usage{
\S4method{runOmicsPCA}{RflomicsSE}(object, ncomp = 5, raw = FALSE)

\S4method{runOmicsPCA}{RflomicsMAE}(object, SE.name, ncomp = 5, raw = FALSE)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE-class}.}

\item{ncomp}{Number of components to compute. Default is 5.}

\item{raw}{boolean. Does the pca have to be ran on raw data or transformed 
and normalized data? Default is FALSE, pca is ran on transformed and 
normalized data.}

\item{SE.name}{the name of the data the normalization have to be applied to.}
}
\value{
An object of class \link{RflomicsSE}
}
\description{
This function performs a principal component analysis on omic 
data stored in an object of class \link{RflomicsSE-class}
Results are stored in the metadata slot of the same object. If a 
"Normalization" slot is present in the metadata slot, then data are 
normalized before running the PCA according to the indicated transform 
method.
}
