% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\name{RunPCA,MultiAssayExperiment-method}
\alias{RunPCA,MultiAssayExperiment-method}
\title{RunPCA}
\usage{
\S4method{RunPCA}{MultiAssayExperiment}(object, data, PCA)
}
\arguments{
\item{data}{omic data type}

\item{PCA}{norm or raw}
}
\value{
MultiAssayExperiment
}
\description{
RunPCA
}
\examples{

}
