% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRFLOMICS.R
\name{runRFLOMICS}
\alias{runRFLOMICS}
\title{runRFLOMICS}
\usage{
runRFLOMICS(...)
}
\arguments{
\item{...}{More arguments to pass to shinyApp.}
}
\value{
shinyApp
}
\description{
running this function will open the shiny application. 
Run the shiny application
}
\examples{
library(RFLOMICS)
runRFLOMICS()


}
