% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_03_data_explor.R
\name{runSampleFiltering,RflomicsSE-method}
\alias{runSampleFiltering,RflomicsSE-method}
\alias{runSampleFiltering}
\title{runSampleFiltering}
\usage{
\S4method{runSampleFiltering}{RflomicsSE}(object, samples = NULL)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE}}

\item{samples}{samples to keep.}
}
\value{
An object of class \link{RflomicsSE}
}
\description{
This function applied sample filtering on an omic data sets stored in an object of
class \link{RflomicsSE}.
}
\examples{
#' Set the data path
datPath <- paste0(system.file(package = "RFLOMICS"), "/ExamplesFiles/ecoseed/")

# Load the experimental design
ExpDesign <- readExpDesign(file = paste0(datPath, "condition.txt"))

# Set factor name, levels, level of reference and type
facRef <- data.frame(factorName   = c("Repeat", "temperature" , "imbibition"), 
                     factorRef = c("rep1", "Low", "DS"), 
                     factorType = c("batch",  "Bio", "Bio"), 
                     factorLevels = c("rep1,rep2,rep3","Low,Medium,Elevated","DS,EI,LI"))

# Load the omics data
omicsData <- list( 
readOmicsData(file = paste0( datPath, "transcriptome_ecoseed.txt")), 
readOmicsData(file = paste0(datPath, "proteome_ecoseed.txt")))
 
# Instantiate an object of class RflomicsMAE 
MAE <- createRflomicsMAE( projectName = "Tests", omicsData   = omicsData,
                         omicsNames  = c("RNAtest.raw", "protetest.raw"),
                         omicsTypes  = c("RNAseq","proteomics"),
                         ExpDesign   = ExpDesign, factorRef   = facRef)
names(MAE) <- c("RNAtest.raw", "protetest.raw")

# Set the statistical model and contrasts to test
formulae <- generateModelFormulae(MAE)
MAE <- setModelFormula(MAE, formulae[[1]])  

# Get the contrasts List and choose the first 3 contrasts of type averaged
contrastList <- getPossibleContrasts(MAE, formula = formulae[[1]], 
                                     typeContrast = "averaged", 
                                     returnTable = TRUE)[c(1, 2, 3),]

# Run the data preprocessing and perform the differential analysis
SE.RNA <- runSampleFiltering(MAE[[1]],samples=colnames(MAE[[1]])[1:6])
SE.prot <- runSampleFiltering(MAE[[2]],samples=colnames(MAE[[2]])[1:6])
}
