% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_03_data_explor.R
\name{runTransformData,RflomicsSE-method}
\alias{runTransformData,RflomicsSE-method}
\alias{runTransformData}
\alias{runTransformData,RflomicsMAE-method}
\title{runTransformData}
\usage{
\S4method{runTransformData}{RflomicsSE}(object, transformMethod = NULL, modifyAssay = FALSE)

\S4method{runTransformData}{RflomicsMAE}(object, SE.name, transformMethod = NULL, modifyAssay = FALSE)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE}}

\item{transformMethod}{The transformation to store in the metadata or to 
store and apply if modifyAssay is TRUE.}

\item{modifyAssay}{Boolean. Do the transformation need to be applied on 
the data? The raw data will be replaced by the transformed ones.}

\item{SE.name}{the name of the data the normalization have to be applied to.}
}
\value{
An object of class \link{RflomicsSE}
}
\description{
runTransformData

runTransformData
}
\examples{

# Set the data path
datPath <- paste0(system.file(package = "RFLOMICS"), "/ExamplesFiles/ecoseed/")

# Load the experimental design
ExpDesign <- readExpDesign(file = paste0(datPath, "condition.txt"))

# Set factor name, levels, level of reference and type
facRef <- data.frame(factorName   = c("Repeat", "temperature" , "imbibition"), 
                     factorRef = c("rep1", "Low", "DS"), 
                     factorType = c("batch",  "Bio", "Bio"), 
                     factorLevels = c("rep1,rep2,rep3","Low,Medium,Elevated","DS,EI,LI"))

# Load the omics data
omicsData <- list( 
readOmicsData(file = paste0(datPath, "proteome_ecoseed.txt")))
 
# Instantiate an object of class RflomicsMAE 
MAE <- createRflomicsMAE( projectName = "Tests", omicsData   = omicsData,
                         omicsNames  = c("protetest"),omicsTypes  = c("proteomics"),
                         ExpDesign   = ExpDesign, factorRef   = facRef)
names(MAE) <- c("protetest")

# Set the statistical model and contrasts to test
formulae <- generateModelFormulae(MAE)
MAE <- setModelFormula(MAE, formulae[[1]])  

# Get the contrasts List and choose the first 3 contrasts of type averaged
contrastList <- getPossibleContrasts(MAE, formula = formulae[[1]], 
                                     typeContrast = "averaged", 
                                     returnTable = TRUE)[c(1, 2, 3),]

# Run the data preprocessing and perform the differential analysis
 MAE |>  runTransformData(SE.name = "protetest",  transformMethod = "log2")
 
 # Get the transformation settings
 getTransSettings(MAE,SE.name = "protetest")
 
}
