% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_03_data_explor.R
\name{setCoeffNorm,RflomicsSE-method}
\alias{setCoeffNorm,RflomicsSE-method}
\alias{setCoeffNorm}
\alias{setCoeffNorm,RflomicsMAE-method}
\title{set Normalization coefficient}
\usage{
\S4method{setCoeffNorm}{RflomicsSE}(object, coeff = NULL)

\S4method{setCoeffNorm}{RflomicsMAE}(object, SE.name = NULL, coeff = NULL)
}
\arguments{
\item{object}{of class RflomicsSE}

\item{coeff}{normalization coefficient}

\item{SE.name}{the name of the data to fetch in the object if the object is a RflomicsMAE}
}
\value{
object of class RflomicsSE
}
\description{
set Normalization coefficient

setCoeffNorm
}
