% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_06_annot_enrichment.R
\name{setEnrichNull,RflomicsSE-method}
\alias{setEnrichNull,RflomicsSE-method}
\alias{setEnrichNull}
\title{Set enrichment results to null
Users should not have to use this method. It is mainly planned for the
interface code.}
\usage{
\S4method{setEnrichNull}{RflomicsSE}(object, from = "DiffExp", database = NULL)
}
\arguments{
\item{object}{a RflomicsSE object}

\item{from}{where to search for the results (either coexp or diffExp)}

\item{database}{which database (GO, KEGG, custom...)}
}
\value{
an RflomicsSE object with NULL value in the
enrichment results.
}
\description{
Set enrichment results to null
Users should not have to use this method. It is mainly planned for the
interface code.
}
\examples{
# Generate RflomicsMAE for example
MAEtest <- generateExample(annotation = FALSE, integration = FALSE)
# Run KEGG annotation on differential analysis results
MAEtest[["protetest"]] <- runAnnotationEnrichment(MAEtest[["protetest"]],
                        list_args = list(organism = "ath",
                                         keyType = "kegg",
                                         pvalueCutoff = 0.1),
                        from = "DiffExp", database = "KEGG")
# Search for the pvalue cutoff:
setEnrichNull(MAEtest[["protetest"]], from = "diffexp", database = "KEGG")

}
