% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_02_set_stat_model.R
\name{setModelFormula,RflomicsMAE-method}
\alias{setModelFormula,RflomicsMAE-method}
\alias{setModelFormula}
\alias{setModelFormula,RflomicsSE-method}
\title{setModelFormula}
\usage{
\S4method{setModelFormula}{RflomicsMAE}(object, modelFormula = NULL)

\S4method{setModelFormula}{RflomicsSE}(object, modelFormula = NULL)
}
\arguments{
\item{object}{a RflomicsSE}

\item{modelFormula}{a formula}
}
\value{
a object RflomicsMAE

a object RflomicsSE
}
\description{
Set model formula from a Flomics RflomicsMAE
}
\seealso{
\link{RflomicsMAE-methods}
}
