% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\name{SetModelMatrix,ExpDesign-method}
\alias{SetModelMatrix,ExpDesign-method}
\title{SetModelMatrix}
\usage{
\S4method{SetModelMatrix}{ExpDesign}(object)
}
\arguments{
\item{ExpDesign}{}
}
\value{

}
\description{
SetModelMatrix
}
