% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\name{SetModelMatrixAndContrasts,ExpDesign-method}
\alias{SetModelMatrixAndContrasts,ExpDesign-method}
\title{SetModelMatrixAndContrasts
Function from DicoExpress}
\usage{
\S4method{SetModelMatrixAndContrasts}{ExpDesign}(object)
}
\arguments{
\item{ExpDesign}{}
}
\value{

}
\description{
SetModelMatrixAndContrasts
Function from DicoExpress
}
\examples{

}
