% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_03_data_explor.R
\name{setNorm,RflomicsSE-method}
\alias{setNorm,RflomicsSE-method}
\alias{setNorm}
\alias{setNorm,RflomicsMAE-method}
\title{set Normalization  method}
\usage{
\S4method{setNorm}{RflomicsSE}(object, method = "none")

\S4method{setNorm}{RflomicsMAE}(object, SE.name = NULL, method = "none")
}
\arguments{
\item{object}{of class RflomicsSE}

\item{method}{the Normalization method to set}

\item{SE.name}{the name of the data to fetch in the object if the object 
is a RflomicsMAE}
}
\value{
object of class RflomicsSE

object of class RflomicsMAE
}
\description{
set Normalization  method

setNorm
}
