% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_02_set_stat_model.R
\name{setSelectedContrasts,RflomicsMAE-method}
\alias{setSelectedContrasts,RflomicsMAE-method}
\alias{setSelectedContrasts}
\alias{setSelectedContrasts,RflomicsSE-method}
\title{setSelectedContrasts}
\usage{
\S4method{setSelectedContrasts}{RflomicsMAE}(object, contrastList = NULL)

\S4method{setSelectedContrasts}{RflomicsSE}(object, contrastList = NULL)
}
\arguments{
\item{object}{a RflomicsSE}

\item{contrastList}{a data.frame of contrast}
}
\value{
a RflomicsMAE object

a RflomicsSE object
}
\description{
A short description...

A short description...
}
