% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_03_data_explor.R
\name{setTrans,RflomicsSE-method}
\alias{setTrans,RflomicsSE-method}
\alias{setTrans}
\alias{setTrans,RflomicsMAE-method}
\title{set transformation  method}
\usage{
\S4method{setTrans}{RflomicsSE}(object, method = "none")

\S4method{setTrans}{RflomicsMAE}(object, SE.name, method = "none")
}
\arguments{
\item{object}{of class RflomicsSE}

\item{method}{the transformation method to set}

\item{SE.name}{the name of the data to fetch in the object if the object 
is a RflomicsMAE}
}
\value{
a RflomicsSE or RflomicsMAE object.
}
\description{
set transformation  method

set transformation  method

setTrans
}
