% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_02_set_stat_model.R
\name{setValidContrasts,RflomicsMAE-method}
\alias{setValidContrasts,RflomicsMAE-method}
\alias{setValidContrasts}
\alias{setValidContrasts,RflomicsSE-method}
\title{setValidContrasts}
\usage{
\S4method{setValidContrasts}{RflomicsMAE}(object, omicName, contrastList = NULL)

\S4method{setValidContrasts}{RflomicsSE}(object, contrastList = NULL)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE-class}}

\item{omicName}{a dataset name}

\item{contrastList}{A data.frame of contrast}
}
\value{
An object of class \link{RflomicsMAE-class}

An object of class \link{RflomicsSE-class}
}
\description{
Set Valid Contrasts

Set Valid Contrasts
}
