% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_04_diff_analysis.R
\name{sumDiffExp}
\alias{sumDiffExp}
\title{Get summary table from diffExpAnalysis analysis}
\usage{
sumDiffExp(object, SE.name = NULL)
}
\arguments{
\item{object}{a SE object (produced by Flomics) or a MAE}

\item{SE.name}{The name of the RflomicsSE to access if object is RflomicsMAE.}
}
\value{
a table
}
\description{
Get summary table from diffExpAnalysis analysis
}
