% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_06_annot_enrichment.R
\name{sumORA,RflomicsSE-method}
\alias{sumORA,RflomicsSE-method}
\alias{sumORA}
\title{Get summary tables from ORA analyses -
once an enrichment has been conducted.}
\usage{
\S4method{sumORA}{RflomicsSE}(
  object,
  from = "DiffExpEnrichAnal",
  database = NULL,
  contrastName = NULL
)
}
\arguments{
\item{object}{a SE object (produced by Flomics)}

\item{from}{either DiffExpEnrichAnal or CoExpAnal.}

\item{database}{either NULL, GO, KEGG or custom.
if NULL, all tables are returned in a list.}

\item{contrastName}{the contrastName or clusterName to retrieve 
the results from. If NULL, all results are returned.}
}
\value{
a list of tables or a table
}
\description{
Get summary tables from ORA analyses -
once an enrichment has been conducted.
}
\examples{
# Generate RflomicsMAE for example
MAEtest <- generateExample(annotation = FALSE, integration = FALSE)
# Run KEGG annotation on differential analysis results
MAEtest[["protetest"]] <- runAnnotationEnrichment(MAEtest[["protetest"]],
                        list_args = list(organism = "ath",
                                         keyType = "kegg",
                                         pvalueCutoff = 0.1),
                        from = "DiffExp", database = "KEGG")
# Search for the pvalue cutoff:
sumORA(MAEtest[["protetest"]], from = "diffexp", database = "KEGG")

}
