% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_01_load_data.R
\name{createRflomicsMAE}
\alias{createRflomicsMAE}
\title{createRflomicsMAE}
\usage{
createRflomicsMAE(
  projectName = NULL,
  omicsData = NULL,
  omicsNames = NULL,
  omicsTypes = NULL,
  ExpDesign = NULL,
  factorRef = NULL
)
}
\arguments{
\item{projectName}{Project name}

\item{omicsData}{list of omics dataset.}

\item{omicsNames}{vector of dataset names.}

\item{omicsTypes}{vector of dataset types: "RNAseq", "metabolomics", "proteomics"}

\item{ExpDesign}{a data.frame imported by \link{readExpDesign} which describes the experimental design.}

\item{factorRef}{a data.frame describing the experimental factors.
\itemize{
\item{factorName:}{factor names}
\item{factorRef:}{factor references}
\item{factorType:}{factor type : "Bio", "batch", "Meta"}
\item{factorLevels:}{levels of each factor with "," separation.}
}}
}
\value{
An object of class \link{RflomicsMAE}
}
\description{
This function is a constructor for the class \link{RflomicsMAE}.
It initializes an object of class \link{RflomicsMAE} from a list of omics datasets
imported by \link{readOmicsData}, a vector of dataset names, a vector of omics types, and 
an experimental design
}
\examples{
# Set the data path
datPath <- paste0(system.file(package = "RFLOMICS"), "/ExamplesFiles/ecoseed/")
ExpDesign <- readExpDesign(file = paste0(datPath, "condition.txt"))

factorRef <- data.frame(factorName  = c("Repeat", "temperature" , "imbibition"),
factorRef   = c("rep1",   "Low",          "DS"),
factorType  = c("batch",  "Bio",          "Bio"),
factorLevels= c("rep1,rep2,rep3", "Low,Medium,Elevated", "DS,EI,LI"))

omicsData <- list(
  RFLOMICS::readOmicsData(file = paste0(datPath, "transcriptome_ecoseed.txt")),
  RFLOMICS::readOmicsData(file = paste0(datPath, "metabolome_ecoseed.txt")),
  RFLOMICS::readOmicsData(file = paste0(datPath, "proteome_ecoseed.txt")))

MAE <- RFLOMICS::createRflomicsMAE(projectName = "Example",
                                               omicsData   = omicsData,
                                               omicsNames  = c("RNAtest", "metatest", "protetest"),
                                               omicsTypes  = c("RNAseq","metabolomics","proteomics"),
                                               ExpDesign   = ExpDesign,
                                               factorRef   = factorRef)
}
\seealso{
\link{RflomicsMAE-class}
}
