% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Class.R
\docType{class}
\name{ExpDesign-class}
\alias{ExpDesign-class}
\alias{.ExpDesign}
\title{ExpDesign Class}
\value{
ExpDesign object
}
\description{
ExpDesign Class
}
\section{Slots}{

\describe{
\item{\code{List.Factors}}{A list of factor}

\item{\code{Factors.Type}}{Either 'Biological' or 'batch'}

\item{\code{Groups}}{design groups}

\item{\code{Model.formula}}{Modele formula.}

\item{\code{Model.matrix}}{GLM model}

\item{\code{Contrasts.List}}{A list of contrasts.}

\item{\code{Contrasts.Sel}}{selected contrast}

\item{\code{Contrasts.Coeff}}{contrast vector}
}}

