% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\name{ExpDesign-Constructor}
\alias{ExpDesign-Constructor}
\alias{ExpDesign.constructor}
\title{[\code{\link{ExpDesign-class}}] Class constructor}
\usage{
ExpDesign.constructor(ExpDesign, projectName, refList, typeList)
}
\arguments{
\item{ExpDesign}{data.frame with experimental design}

\item{refList}{vector with factor ref}

\item{typeList}{vector with type of factor}
}
\value{
An object of class [\code{\link{ExpDesign-class}}]
}
\description{
[\code{\link{ExpDesign-class}}] Class constructor
}
