% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_02_set_stat_model.R
\name{generateExpressionContrast,RflomicsSE-method}
\alias{generateExpressionContrast,RflomicsSE-method}
\alias{generateExpressionContrast}
\alias{generateExpressionContrast,RflomicsMAE-method}
\title{Get model formula from a Flomics RflomicsSE}
\usage{
\S4method{generateExpressionContrast}{RflomicsSE}(object)

\S4method{generateExpressionContrast}{RflomicsMAE}(object)
}
\arguments{
\item{object}{An object of class [\code{\link{RflomicsSE-class}}] 
or class [\code{\link{RflomicsMAE-class}}]}
}
\value{
list of 1 or 3 data.frames of contrast expression

list of 1 or 3 data.frames of contrast expression
}
\description{
This function allows, from a model formulae, to give the 
expression contrast data frames.
Three types of contrasts are expressed:
\itemize{
\item{pairwise comparison}
\item{averaged expression}
\item{interaction expression}
}

This function allows, from a model formulae, 
to give the expression contrast data frames.
Three types of contrasts are expressed:
\itemize{
\item{pairwise comparison}
\item{averaged expression}
\item{interaction expression}
}
}
\author{
Christine Paysant-Le Roux, adapted by Nadia Bessoltane
}
