% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_00_Common.R
\name{generateReport,RflomicsMAE-method}
\alias{generateReport,RflomicsMAE-method}
\alias{generateReport}
\title{Generate RFLOMICS rmarkdown report}
\usage{
\S4method{generateReport}{RflomicsMAE}(
  object,
  fileName = NULL,
  archiveName = NULL,
  export = FALSE,
  tmpDir = getwd(),
  ...
)
}
\arguments{
\item{object}{a \link{RflomicsMAE} produced by RFLOMICS.}

\item{fileName}{Name of the html report (default: date()_projectName.html).}

\item{archiveName}{name of archive with all analysis results 
(default: date()_projectName.tar.gz).}

\item{export}{boolean value to create archive (default: FALSE)}

\item{tmpDir}{temporary directory (default: working directory)}

\item{...}{other arguments to pass into the render function.}
}
\value{
An html report or archive (tar.gz)
}
\description{
This function is used to generate a html report from a
\link{RflomicsMAE} object or archive with results.
}
\examples{
ExpDesign <- RFLOMICS::readExpDesign(
  file = paste0(system.file(package = "RFLOMICS"), 
                "/ExamplesFiles/ecoseed/condition.txt"))
factorRef <- data.frame(
  factorName  = c("Repeat", "temperature" , "imbibition"),
  factorRef   = c("rep1",   "Low",          "DS"),
  factorType  = c("batch",  "Bio",          "Bio"),
  factorLevels= c("rep1,rep2,rep3", "Low,Medium,Elevated", "DS,EI,LI"))

omicsData <- list(
  RFLOMICS::readOmicsData(
    file = paste0(system.file(package = "RFLOMICS"), 
                  "/ExamplesFiles/ecoseed/proteome_ecoseed.txt")))

omicsNames  = c("protetest")
omicsTypes  = c("proteomics")
MAE <- RFLOMICS::createRflomicsMAE(projectName = "Tests",
                                   omicsData   = omicsData,
                                   omicsNames  = omicsNames,
                                   omicsTypes  = omicsTypes,
                                   ExpDesign   = ExpDesign,
                                   factorRef   = factorRef)

formulae <- RFLOMICS::generateModelFormulae(MAE)
MAE <- RFLOMICS::setModelFormula(MAE, modelFormula = formulae[[1]])

Contrasts.List <- RFLOMICS::generateExpressionContrast(MAE)
selectedContrasts <- Contrasts.List$simple[1:4,]

MAE <- setSelectedContrasts(MAE, contrastList = selectedContrasts)

## data processing
MAE <- RFLOMICS::runDataProcessing(
  object = MAE, 
  SE.name = "protetest", 
  samples=NULL, 
  normMethod="none", 
  transformMethod="none")

## diff analysis
MAE <- RFLOMICS::runDiffAnalysis(
  object = MAE, 
  SE.name = "protetest", 
  contrastList = 
    selectedContrasts, 
  p.adj.method="BH", 
  method = "limmalmFit",  
  p.adj.cutoff = 0.05, 
  logFC.cutoff = 0)

## Enrichment
MAE <- RFLOMICS::runAnnotationEnrichment(
  object = MAE, 
  SE.name = "protetest", 
  database = "GO", 
  domain = c("MF"), 
  list_args = list(OrgDb = "org.At.tair.db", 
                   keyType = "TAIR", 
                   pvalueCutoff = 0.05))

#generateReport(object = MAE, fileName = ecoseed_report.html)
}
