% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\name{mvQCdata,MultiAssayExperiment-method}
\alias{mvQCdata,MultiAssayExperiment-method}
\title{multivariate QC data}
\usage{
\S4method{mvQCdata}{MultiAssayExperiment}(object, data, PCA = c("raw", "norm"), axis = 3, pngFile = NULL)
}
\arguments{
\item{data}{data name}

\item{PCA}{pca axis to plot}

\item{axis}{The number of PCA axis}

\item{pngFile}{plot file name to save}

\item{MultiAssayExperiment}{an object of Class MultiAssayExperiment}
}
\description{
multivariate QC data
}
