% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOMICS-Methods.R
\docType{methods}
\name{mvQCdesign,FlomicsExperiment-method}
\alias{mvQCdesign,FlomicsExperiment-method}
\title{Multivariate Quality Check}
\usage{
\S4method{mvQCdesign}{FlomicsExperiment}(object, axis = 3)
}
\arguments{
\item{axis}{The number of PCA axis}

\item{FlomicsExperiment}{an object of Class FlomicsExperiment}
}
\description{
Multivariate Quality Check
}
