% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Class.R
\docType{class}
\name{RflomicsMAE-class}
\alias{RflomicsMAE-class}
\title{\link{RflomicsMAE-class} Class}
\value{
A \code{RflomicsMAE} object.
}
\description{
RflomicsMAE is a class that extends the \link{MultiAssayExperiment}  class by imposing a structure
to the metadata slot. This class is used by the Rflomics analysis workflow to store the 
experimental design, the settings and results of a multi-omics integration analysis.  
The slot metadata is structured as follows:
 
 \itemize{
   \item omicList: Contains the list of omics datasets, with the type and name  
   \item projectName: The project name
   \item design: The experimental design 
   \item IntegrationAnalysis: A list containing the multi-omics integration analysis settings and results.   
 }
}
\seealso{
\link{MultiAssayExperiment}

\link{RflomicsMAE-methods}
}
