% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_00_Common.R
\name{rflomicsMAE2MAE}
\alias{rflomicsMAE2MAE}
\alias{rflomicsMAE2MAE,RflomicsMAE-method}
\title{convert rflomicsMAE to MAE}
\usage{
\S4method{rflomicsMAE2MAE}{RflomicsMAE}(object, raw = FALSE)
}
\arguments{
\item{object}{The RflomicsMAE object to convert.}

\item{raw}{booleen. If TRUE raw omics data values.}
}
\value{
object of MultiAssayExperiment class.
}
\description{
Convert rflomicsMAE object to MultiAssayExperiment object.
}
