% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Class.R
\docType{class}
\name{RflomicsSE-class}
\alias{RflomicsSE-class}
\alias{RflomicsSE}
\title{\link{RflomicsSE-class} Class}
\value{
A \code{RflomicsSE} object.
}
\description{
RflomicsSE is a class that extends the \link{SummarizedExperiment} class by imposing a structure
to the metadata slot. This class is used by the Rflomics analysis workflow to store the 
experimental design, the settings and results of a single omic analysis. 
The slot metadata is structured as follows:

\itemize{
  \item omicType: the type of omics dataset
  \item design: experimental design
  \item DataProcessing: a list of data processing settings and results 
  \item PCAlist: a list of PCA settings and results
  \item DiffExpAnal: a list containing the Differential Analysis settings and results
  \item CoExpAnal: a list containing the Coexpression Analysis settings and results
  \item DiffExpEnrichAnal: a list containing the enrichment analysis of the list of DE features settings and results
  \item CoExpEnrichAnal: a list containing the enrichment analysis of the list of co-expressed features settings and results
 }
}
\seealso{
\link{SummarizedExperiment}

\link{RflomicsSE-methods}
}
