% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_06_annot_enrichment.R
\name{runAnnotationEnrichment,RflomicsSE-method}
\alias{runAnnotationEnrichment,RflomicsSE-method}
\alias{runAnnotationEnrichment}
\alias{runAnnotationEnrichment,RflomicsMAE-method}
\title{runAnnotationEnrichment}
\usage{
\S4method{runAnnotationEnrichment}{RflomicsSE}(
  object,
  nameList = NULL,
  list_args = list(pvalueCutoff = 0.05, qvalueCutoff = 1, minGSSize = 10, maxGSSize =
    500, universe = names(object)),
  from = "DiffExp",
  database = "custom",
  domain = "no-domain",
  col_term = "term",
  col_gene = "gene",
  col_name = "name",
  col_domain = NULL,
  annot = NULL
)

\S4method{runAnnotationEnrichment}{RflomicsMAE}(
  object,
  SE.name,
  nameList = NULL,
  list_args = list(pvalueCutoff = 0.05, qvalueCutoff = 1, minGSSize = 10, maxGSSize =
    500, universe = names(object)),
  from = "DiffExp",
  database = "custom",
  domain = "no-domain",
  col_term = "term",
  col_gene = "gene",
  col_name = "name",
  col_domain = NULL,
  annot = NULL
)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE} or
\link{RflomicsMAE}. It is expected the SE object is produced by
rflomics previous analyses, as it relies on their results.}

\item{nameList}{name of contrasts (tags or names) from which to extract DE
genes if from is DiffExpAnal.}

\item{list_args}{list of arguments to pass to the enrichment function.
These arguments must match the ones from the clusterprofiler package.
E.g: universe, keytype, pvalueCutoff, qvalueCutoff, etc.}

\item{from}{indicates if ListNames are from differential analysis results
(DiffExpAnal) or from the co-expression analysis results (CoExpAnal)}

\item{database}{is it a custom annotation, GO or KEGG annotations}

\item{domain}{subcatgory for the database (eg BP for GO)}

\item{col_term}{for custom annotation, column name of the file containing
term names or id.}

\item{col_gene}{for custom annotation, column name of the file containing
entities names.}

\item{col_name}{(optional) for custom annotation, column name of the 
file containing term names if col_term is used for ids.}

\item{col_domain}{for custom annotation, column name of the file containing
the domains.}

\item{annot}{for custom annotation, path of the annotation file.}

\item{SE.name}{name of the experiment to consider if object is a
RflomicsMAE.}
}
\value{
A RflomicsMAE or a RflomicsS, depending on the class of object
parameter. The enrichment results are added to the metadata slot, either
in DiffExpEnrichAnal or CoExpEnrichAnal.
}
\description{
This function performs overrepresentation analysis (ORA) using
clusterprofiler functions. It can be used with custom annotation file
(via enricher), GO (enrichGO) or KEGG (enrichKEGG) annotations.
}
\examples{
# Generate RflomicsMAE for example
 datPath <- paste0(system.file(package = "RFLOMICS"),
                   "/ExamplesFiles/ecoseed/")

ExpDesign <- readExpDesign(file = paste0(datPath, "condition.txt"))
facRef <- data.frame(
                factorName   = c("Repeat", "temperature" , "imbibition"),
                factorRef    = c("rep1",   "Low",          "DS"),
                factorType   = c("batch",  "Bio",          "Bio"),
                factorLevels = c("rep1,rep2,rep3",
                                 "Low,Medium,Elevated",
                                 "DS,EI,LI"))
                       
omicsData <- list(
   readOmicsData(file = paste0(datPath, "transcriptome_ecoseed.txt")),
   readOmicsData(file = paste0(datPath, "metabolome_ecoseed.txt")))
   
MAEtest <- createRflomicsMAE(projectName = "Tests",
                             omicsData   = omicsData,
                             omicsNames  = c("RNAtest",  "protetest"),
                             omicsTypes  = c("RNAseq", "proteomics"),
                             ExpDesign   = ExpDesign,
                             factorRef   = facRef
                             )
names(MAEtest) <- c("RNAtest", "protetest")

# Run GO annotation (enrichGO)
MAEtest[["protetest"]] <- runAnnotationEnrichment(MAEtest[["protetest"]],
                        list_args = list(OrgDb = "org.At.tair.db",
                                         keyType = "TAIR",
                                         pvalueCutoff = 0.05),
                        from = "DiffExp", database = "GO",
                        domain = "CC")
getEnrichRes(MAEtest[["protetest"]])

# Run KEGG annotation (enrichKEGG)
MAEtest[["protetest"]] <- runAnnotationEnrichment(MAEtest[["protetest"]],
                        list_args = list(organism = "ath",
                                         keyType = "kegg",
                                         pvalueCutoff = 0.05),
                        from = "DiffExp", database = "KEGG")
sumORA(MAEtest[["protetest"]], from = "DiffExp", database = "KEGG")
}
