% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_05_coexp_analysis.R
\name{runCoExpression,RflomicsSE-method}
\alias{runCoExpression,RflomicsSE-method}
\alias{runCoExpression}
\alias{runCoExpression,RflomicsMAE-method}
\title{runCoExpression}
\usage{
\S4method{runCoExpression}{RflomicsSE}(
  object,
  K = 2:20,
  replicates = 5,
  contrastNames = NULL,
  merge = "union",
  model = "Normal",
  GaussianModel = NULL,
  transformation = NULL,
  normFactors = NULL,
  clustermq = FALSE,
  meanFilterCutoff = NULL,
  scale = NULL,
  silent = TRUE,
  cmd = FALSE
)

\S4method{runCoExpression}{RflomicsMAE}(
  object,
  SE.name,
  K = 2:20,
  replicates = 5,
  contrastNames,
  merge = "union",
  model = "Normal",
  GaussianModel = NULL,
  transformation = NULL,
  normFactors = NULL,
  clustermq = FALSE,
  meanFilterCutoff = NULL,
  scale = NULL,
  silent = TRUE,
  cmd = FALSE
)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE}}

\item{K}{Number of clusters (a single value or a vector of values)}

\item{replicates}{The number of iteration for each K.}

\item{contrastNames}{names of the contrasts from which the DE entities 
have to be taken. Can be NULL, in that case every contrasts from the differential 
analysis are taken into consideration.}

\item{merge}{\code{"union"} or \code{"intersection"}}

\item{model}{Type of mixture model to use \code{"Poisson"} or 
\code{"normal"}. By default, it is the normal.}

\item{GaussianModel}{Type of \code{GaussianModel} to be used for the 
Normal mixture model only. This parameters
is set to \code{"Gaussian_pk_Lk_Ck"} by default and doesn't have to 
be changed except if an error message proposed
to try another model like \code{"Gaussian_pk_Lk_Bk"}.}

\item{transformation}{The transformation type to be used. By default, 
it is the "arcsin" one.}

\item{normFactors}{The type of estimator to be used to normalize for 
differences in library size.
By default, it is the "TMM" one.}

\item{clustermq}{boolean. Does the computation need to be executed 
on a distant server. A configuration file and a network connection are 
needed to use this option.}

\item{meanFilterCutoff}{a cutoff to filter a gene with a mean expression lower than this value.
(only for RNAseq data, set to NULL for others).}

\item{scale}{Boolean. If TRUE scale all variables tounit variance.}

\item{silent}{if TRUE, coseq run silently (without any console print or 
message)}

\item{cmd}{if TRUE, print steps of the analysis. Used inside the coseq 
module in the shiny interface.}

\item{SE.name}{the name of the data to fetch in the object if the object 
is a RflomicsMAE}
}
\value{
An S4 object of class \link{RflomicsSE}
All the results are stored as a named list \code{CoExpAnal}
 in the metadata slot of a given \code{RflomicsSE} object. Objects are:
The runCoExpression method return several results, for \link{coseq} 
method, objects are:
\itemize{
\item{\code{setting:} }{co-expression analysis settings. See \code{getCoexpSetting}}
\item{\code{results:} }{boolean indicating if the co-expression analysis succeed}
\item{\code{coseqResults:} }{the raw results of \code{coseq}}
\item{\code{clusters:} }{a List of clusters}
\item{\code{cluster.nb:} }{The number of cluster}
\item{\code{plots:} }{The plots of \code{coseq} results}
\item{\code{stats:} }{A tibble summarising failed jobs: reason, propoif any}
}
}
\description{
This method performs a co-expression/co-abundance analysis of 
omic-data.
}
\details{
For now, only the coseq function of the coseq package is used.
For RNAseq data, parameters used are those recommended in DiCoExpress 
workflow (see the reference). This parameters are: \code{model="normal"}, 
\code{transformation="arcsin"}, \code{GaussianModel="Gaussian_pk_Lk_Ck"},
\code{normFactors="TMM"}, \code{meanFilterCutoff = 50}
For proteomic or metabolomic, data are scaled by protein or metabolite 
to group them by expression profiles rather than by expression intensity.
After data scaling, recommended parameters (from \code{coseq} developers) 
for co-expression analysis are:
\code{model="normal"}, \code{transformation="none"}, 
\code{GaussianModel="Gaussian_pk_Lk_Ck"},
\code{normFactors="none"},  \code{meanFilterCutoff = NULL}.
}
\examples{

# Set the data path
datPath <- paste0(system.file(package = "RFLOMICS"), "/ExamplesFiles/ecoseed/")

# Load the experimental design
ExpDesign <- readExpDesign(file = paste0(datPath, "condition.txt"))

# Set factor name, levels, level of reference and type
facRef <- data.frame(factorName   = c("Repeat", "temperature" , "imbibition"), 
                     factorRef = c("rep1", "Low", "DS"), 
                     factorType = c("batch",  "Bio", "Bio"), 
                     factorLevels = c("rep1,rep2,rep3","Low,Medium,Elevated","DS,EI,LI"))

# Load the omics data
omicsData <- list( 
readOmicsData(file = paste0(datPath, "proteome_ecoseed.txt")))
 
# Instantiate an object of class RflomicsMAE 
MAE <- createRflomicsMAE( projectName = "Tests", omicsData   = omicsData,
                         omicsNames  = c("protetest"), omicsTypes  = c("proteomics"),
                         ExpDesign   = ExpDesign, factorRef   = facRef)
names(MAE) <- c("protetest")

# Set the statistical model and contrasts to test
formulae <- generateModelFormulae(MAE)
MAE <- setModelFormula(MAE, formulae[[1]])  

# Get the contrasts List and choose the first 3 contrasts of type averaged
contrastList <- getPossibleContrasts(MAE, 
formula = formulae[[1]], 
                                     typeContrast = "averaged", 
                                     returnTable = TRUE)[c(1, 2, 3),]

# Run the data preprocessing and perform the differential analysis
 MAE <- MAE |>  runTransformData(SE.name = "protetest",  transformMethod = "log2") |>
 runNormalization(SE.name = "protetest", normMethod = "median")    |> 
 runDiffAnalysis(SE.name = "protetest",  
                 method = "limmalmFit", contrastList = contrastList) |> 
 runCoExpression(SE.name = "protetest", K = 2:5, replicates = 5, 
                 contrastNames = contrastList$tag, merge = "union")
 
 getCoexpSettings(MAE[["protetest"]])
 
 clusters <- getCoseqClusters(MAE[["protetest"]])
 
}
\references{
Lambert, I., Paysant-Le Roux, C., Colella, S. et al. DiCoExpress: 
a tool to process multifactorial RNAseq experiments from quality controls 
to co-expression analysis through differential analysis based on contrasts 
inside GLM models. Plant Methods 16, 68 (2020).
}
\seealso{
\code{\link[coseq]{coseq}}
}
