% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_03_data_explor.R
\name{runDataProcessing,RflomicsSE-method}
\alias{runDataProcessing,RflomicsSE-method}
\alias{runDataProcessing}
\alias{runDataProcessing,RflomicsMAE-method}
\title{runDataProcessing}
\usage{
\S4method{runDataProcessing}{RflomicsSE}(
  object,
  samples = NULL,
  lowCountFiltering_strategy = "NbReplicates",
  lowCountFiltering_CPM_Cutoff = 1,
  normMethod = "none",
  transformMethod = "none"
)

\S4method{runDataProcessing}{RflomicsMAE}(
  object,
  samples = NULL,
  lowCountFiltering_strategy = "NbReplicates",
  lowCountFiltering_CPM_Cutoff = 1,
  normMethod = "none",
  transformMethod = "none",
  SE.name
)
}
\arguments{
\item{object}{An object of class \link{RflomicsSE}}

\item{samples}{samples to keep.}

\item{lowCountFiltering_strategy}{strategy of RNAseq low count filtering. 
Mandatory for RNAseq data. Default value : "NbReplicates".}

\item{lowCountFiltering_CPM_Cutoff}{CPM cutoff for RNAseq low count filtering.
Mandatory for RNAseq data. Default value : 1.}

\item{normMethod}{of normalisation. Mandatory for RNAseq data. 
Default value : RNAseq = TMM.}

\item{transformMethod}{method of transformation.}

\item{SE.name}{the name of the data the normalization have to be applied to.}
}
\value{
An object of class \link{RflomicsSE}
}
\description{
This function applied a processing (filtering, normalization and/or transformation, PCA) on an omic data sets stored in an object of
class \link{RflomicsSE}.
\itemize{
\item{For RNAseq data:}{}
\item{For Proteomic data:}{}
\item{For Metabolomic data:}{}
}
}
\examples{

# Set the data path
datPath <- paste0(system.file(package = "RFLOMICS"), "/ExamplesFiles/ecoseed/")

# Load the experimental design
ExpDesign <- readExpDesign(file = paste0(datPath, "condition.txt"))

# Set factor name, levels, level of reference and type
facRef <- data.frame(factorName   = c("Repeat", "temperature" , "imbibition"), 
                     factorRef = c("rep1", "Low", "DS"), 
                     factorType = c("batch",  "Bio", "Bio"), 
                     factorLevels = c("rep1,rep2,rep3","Low,Medium,Elevated","DS,EI,LI"))

# Load the omics data
omicsData <- list( 
readOmicsData(file = paste0( datPath, "transcriptome_ecoseed.txt")), 
readOmicsData(file = paste0(datPath, "proteome_ecoseed.txt")))
 
# Instantiate an object of class RflomicsMAE 
MAE <- createRflomicsMAE(projectName = "Tests", omicsData   = omicsData,
                         omicsNames  = c("RNAtest.raw", "protetest.raw"),
                         omicsTypes  = c("RNAseq","proteomics"),
                         ExpDesign   = ExpDesign, factorRef   = facRef)
names(MAE) <- c("RNAtest.raw", "protetest.raw")

# Set the statistical model and contrasts to test
formulae <- generateModelFormulae(MAE)
MAE <- setModelFormula(MAE, formulae[[1]])  

# Get the contrasts List and choose the first 3 contrasts of type averaged
contrastList <- getPossibleContrasts(MAE, formula = formulae[[1]], 
                                     typeContrast = "averaged", 
                                     returnTable = TRUE)[c(1, 2, 3),]

# Run the data preprocessing and perform the differential analysis
MAE <- runDataProcessing(MAE, SE.name = "RNAtest",  
                         lowCountFiltering_strategy = "NbReplicates", 
                         lowCountFiltering_CPM_Cutoff = 1, 
                         normMethod = "TMM", 
                         transformMethod = "log2") 
MAE <- runDataProcessing(MAE, SE.name = "protetest", 
                         normMethod = "median", 
                         transformMethod = "log2")
}
\seealso{
runSampleFiltering

filterLowAbundance

runNormalization

runTransformData
}
