% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_04_diff_analysis.R
\name{runDiffAnalysis,RflomicsSE-method}
\alias{runDiffAnalysis,RflomicsSE-method}
\alias{runDiffAnalysis}
\alias{runDiffAnalysis,RflomicsMAE-method}
\alias{getDiffSettings,RflomicsSE-method}
\alias{getDiffSettings}
\alias{getDiffSettings,RflomicsMAE-method}
\title{runDiffAnalysis}
\usage{
\S4method{runDiffAnalysis}{RflomicsSE}(
  object,
  p.adj.method = "BH",
  contrastList = NULL,
  method = NULL,
  p.adj.cutoff = 0.05,
  logFC.cutoff = 0,
  clustermq = FALSE,
  parallel = FALSE,
  nworkers = 1,
  cmd = FALSE
)

\S4method{runDiffAnalysis}{RflomicsMAE}(
  object,
  SE.name,
  p.adj.method = "BH",
  contrastList = NULL,
  method = NULL,
  p.adj.cutoff = 0.05,
  logFC.cutoff = 0,
  clustermq = FALSE,
  parallel = FALSE,
  nworkers = 1,
  cmd = FALSE
)

\S4method{getDiffSettings}{RflomicsSE}(object)

\S4method{getDiffSettings}{RflomicsMAE}(object, SE.name)
}
\arguments{
\item{object}{of class RflomicsSE}

\item{p.adj.method}{The method choosen to adjust pvalue. Takes the same 
values as the ones of adj.p.adjust method.}

\item{contrastList}{data.frame of contrast from getExpressionContrastF()}

\item{method}{A character vector giving the name of the differential 
analysis method to run. Either "edgeRglmfit" or "limmalmFit".}

\item{p.adj.cutoff}{The adjusted pvalue cut-off}

\item{logFC.cutoff}{The lof2FC cutoff}

\item{clustermq}{A boolean indicating whether the constrasts have to be 
computed in local or in a distant machine}

\item{parallel}{boolean. If TRUE, uses mclapply function from the parallel
package.}

\item{nworkers}{integers. The number of CPU to use.}

\item{cmd}{boolean. Used in the interface. If TRUE, print cmd for the user.}

\item{SE.name}{the name of the dataset to fetch if the object is 
of class \link{RflomicsMAE}}
}
\value{
All the results are stored as a named list \code{DiffExpAnal} 
in the metadata slot of a given \link{RflomicsSE} object.

Objects are:
\itemize{
\item{stats:}{ data.frame giving a summary of the differential statistical analysis results by contrast:
 number of DE features, number of up and down regulated features}

\item{setting: }{Parameters used for the differential analysis}
\itemize{
\item{method: }{ The method used for the differential analysis}
\item{p.adj.method:}{ The applied p-value correction method}
\item{p.adj.cutoff:}{ The cut-off applied for the adjusted p-value}
\item{logFC.cutoff:}{ The absolute log FC cut-off}
}

\item{RawDEFres: }{a list giving for each contrast the raw results of 
the differential analysis method}
\item{DEF: }{a list giving for each contrast a data.frame of non filtered 
differential expressed features with their statistics}
\item{TopDEF: }{a list giving for each contrast a data.frame of 
differential expressed features ordered and filtered by p.adj.cutoff 
with their statistics}
\item{mergeDEF: }{a data frame of 0/1 indicating for each features in row, 
if it is DE in a given contrasts in column}
\item{contrasts: }{a data.table of the contrasts used for the differential 
analysis}
}

An object of class \link{RflomicsSE}

List of differential analysis setting parametres.
}
\description{
This is an interface method which run a differential 
analysis on omics datasets stored in an object of class 
\link{RflomicsSE} or \link{RflomicsMAE}. According to the type 
of omics and to a list of contrasts, a differential analysis  
is performef for each contrasts.
Two methods are available according to the type of object:
\itemize{
\item{For RNAseq data: }{the \code{glmFit} function/model of the 
\code{edgeR} package} is applied.
\item{For proteomic and metabolomic data: }{the \code{\link{lmFit}} 
function/model of the \code{limma} package} is applied.
}
Functions and parameters used for RNAseq are those recommended in DiCoExpress 
workflow (see the paper in reference).
Functions and parameters used for proteomic and metabolomic data are those 
recommended in the
}
\examples{

# Set the data path
datPath <- paste0(system.file(package = "RFLOMICS"), "/ExamplesFiles/ecoseed/")

# Load the experimental design
ExpDesign <- readExpDesign(file = paste0(datPath, "condition.txt"))

# Set factor name, levels, level of reference and type
facRef <- data.frame(factorName   = c("Repeat", "temperature" , "imbibition"), 
                     factorRef = c("rep1", "Low", "DS"), 
                     factorType = c("batch",  "Bio", "Bio"), 
                     factorLevels = c("rep1,rep2,rep3","Low,Medium,Elevated","DS,EI,LI"))

# Load the omics data
omicsData <- list( 
readOmicsData(file = paste0( datPath, "transcriptome_ecoseed.txt")), 
readOmicsData(file = paste0(datPath, "proteome_ecoseed.txt")))
 
# Instantiate an object of class RflomicsMAE 
MAE <- createRflomicsMAE(projectName = "Tests", omicsData   = omicsData,
                         omicsNames  = c("RNAtest", "protetest"),
                         omicsTypes  = c("RNAseq","proteomics"),
                         ExpDesign   = ExpDesign, factorRef   = facRef)
names(MAE) <- c("RNAtest", "protetest")

# Set the statistical model and contrasts to test
formulae <- generateModelFormulae(MAE)
MAE <- setModelFormula(MAE, formulae[[1]])  

# Get the contrasts List and choose the first 3 contrasts of type averaged
contrastList <- getPossibleContrasts(MAE, formula = formulae[[1]], 
                                     typeContrast = "averaged", 
                                     returnTable = TRUE)[c(1, 2, 3),]

# Run the data preprocessing and perform the differential analysis
MAE <- MAE |>  runTransformData(SE.name = "protetest",  transformMethod = "log2") |>
 filterLowAbundance(SE.name = "RNAtest")                           |>    
 runNormalization(SE.name = "RNAtest",   normMethod = "TMM")       |>
 runNormalization(SE.name = "protetest", normMethod = "median")    |>
 runDiffAnalysis(SE.name = "protetest", method = "limmalmFit", 
                 contrastList = contrastList)  |>
 runDiffAnalysis(SE.name = "RNAtest", method = "edgeRglmfit", 
                 contrastList = contrastList) 
 
 # Access to the diff analysis settings
 getDiffSettings(experiments(MAE)[["RNAtest"]])
 getDiffSettings(experiments(MAE)[["protetest"]])

 getDEList(experiments(MAE)[["RNAtest"]]) 
}
\references{
Lambert, I., Paysant-Le Roux, C., Colella, S. et al. DiCoExpress: a tool 
to process multifactorial RNAseq experiments from quality controls to 
co-expression analysis through differential analysis based on contrasts 
inside GLM models. Plant Methods 16, 68 (2020).
}
\seealso{
\code{\link{filterDiffAnalysis}}

\code{\link{getDEMatrix}}

\code{\link{getDEList}}
}
