% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLOMICS-Methods_07_Integration.R
\name{runOmicsIntegration,RflomicsMAE-method}
\alias{runOmicsIntegration,RflomicsMAE-method}
\alias{runOmicsIntegration}
\title{runOmicsIntegration}
\usage{
\S4method{runOmicsIntegration}{RflomicsMAE}(
  object,
  preparedObject = NULL,
  method = "MOFA",
  scale_views = FALSE,
  maxiter = 1000,
  num_factors = 10,
  selectedResponse = NULL,
  ncomp = 2,
  link_datasets = 1,
  link_response = 1,
  sparsity = FALSE,
  cases_to_try = 5,
  silent = TRUE,
  cmd = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \link{RflomicsMAE}.
It is expected the MAE object is produced by rflomics previous analyses,
as it relies on their results.}

\item{preparedObject}{An untrained MOFA object or a list of dataset.
Usually a result of \link{prepareForIntegration}.}

\item{method}{one of MOFA or mixOmics. 
Method for which the object is prepared.}

\item{scale_views}{boolean. If TRUE, scale each dataset to unit variance.}

\item{maxiter}{MOFA2 parameter. 
Number of max iteration (otherwise stop when converged.)}

\item{num_factors}{MOFA2 parameter. The number of factor to compute.}

\item{selectedResponse}{character vector, used for mixOmics. 
Response variables names for block.(s)plsda.}

\item{ncomp}{mixOmics parameter. Number of components to compute.}

\item{link_datasets}{mixOmics parameter. 
Link between datasets in the computation.}

\item{link_response}{mixOmics parameter. Link between dataset and response.}

\item{sparsity}{boolean. Used to determine which mixOmics function to apply (either
block.plsda if FALSE or block.splsda if TRUE).}

\item{cases_to_try}{integer. If sparsity is set to TRUE, then cases_to_try
is used to determine the number of sets of variables to test for tuning.}

\item{silent}{boolean. If TRUE, silence all functions.}

\item{cmd}{boolean. Used in the interface. If TRUE, print cmd in the console.}

\item{...}{not in use at the moment}
}
\value{
a RflomicsMAE object with the correct metadata slot filled with the
results and the settings.
}
\description{
Runs the integration according to the selected method (MOFA or
mixOmics) and the settings given by the user. Requires to have the correct
entry format in preparedObject before running.
}
\examples{
# Generate MAE for test:
MAEtest <- generateExample(annotation = FALSE, coexp = FALSE,
                           integration = FALSE)
                           
# Prepare mofa object:
mofaObj <- prepareForIntegration(MAEtest,
                omicsNames = c("protetest", "metatest"),
                variableLists = rownames(MAEtest),
                method = "MOFA")

# Perform integration:
MAEtest <- runOmicsIntegration(MAEtest, mofaObj, method = "MOFA")
MOFA2::plot_variance_explained(getMOFA(MAEtest))

}
